json 형식으로 변환 및 코드 분할
