+++ core/file_processor.py
@@ -1,0 +1,20 @@
+import fitz
+from pptx import Presentation
+
+# PDF에서 텍스트 추출
+def extract_text_from_pdf(file_path):
+    text = ""
+    with fitz.open(file_path) as doc:
+        for page in doc:
+            text += page.get_text()
+    return text
+
+# PPTX에서 텍스트 추출
+def extract_text_from_pptx(file_path):
+    text = ""
+    prs = Presentation(file_path)
+    for slide in prs.slides:
+        for shape in slide.shapes:
+            if hasattr(shape, "text"):
+                text += shape.text + "\n"
+    return text
