--- main.py
@@ -1,14 +1,0 @@
-from fastapi import FastAPI, UploadFile, File, Form, HTTPException
-from fastapi.middleware.cors import CORSMiddleware
-
-app = FastAPI()
-
-# CORS 설정 업데이트
-app.add_middleware(
-    CORSMiddleware,
-    allow_origins=["*"],  # 개발 중에는 모든 출처 허용
-    allow_credentials=True,
-    allow_methods=["*"],
-    allow_headers=["*"],
-    expose_headers=["*"],
-)
