+++ prompts/__init__.py
@@ -1,0 +1,9 @@
+from .education_config import user_summary_level, user_question_level, field_features
+from .prompt_manager import get_summary_prompt
+
+__all__ = [
+    'user_summary_level',
+    'user_question_level', 
+    'field_features',
+    'get_summary_prompt'
+]
