+++ prompts/education_config.py
@@ -1,0 +1,67 @@
+## summary_level 수준 정의
+def user_summary_level(summary_level):
+    if summary_level == "비전공자":
+        return f""" 비전공자 수준 요약은 다음과 같은 기준을 따르라.\n
+                [학습자 정의]\n
+                비전공자는 문서에 제시된 정보나 용어를 사실적으로 이해할 수 있으나, 개념 간 관계를 구조화하거나 의미를 확장하는 데에는 어려움을 겪는 학습자이다.\n
+                [학습자 특성]\n
+                -문서에 명시된 정의, 사례, 수치, 설명 등을 표면적으로 요약할 수 있음.
+                -서술된 정보는 항목별로 나열하되, 항목 간의 연결성이나 구조적 흐름은 형성하지 못함.
+                -문서의 중심 개념을 반복 횟수에 따라 중요하게 인식하지만, 그 개념이 수행하는 기능이나 맥락적 의미는 분석하지 않음.
+                -문서 전체의 의도, 전개 방식, 개념적 구조를 통합적으로 파악하지 못함.
+                -전문 용어 사용을 지양하고, 직관적이고 쉬운 표현 위주로 요약을 구성함.\n
+                [요약 수준 예시]\n
+                탄소중립은 온실가스를 줄이거나 흡수해서 실제 배출량을 0으로 만드는 것이다. 
+                이를 위해 교통, 에너지, 건물 등에서 다양한 정책이 시행되고 있다. 우리나라뿐 아니라 여러 나라가 기후변화를 막기 위해 노력하고 있다."""
+
+    elif summary_level == "전공자":
+        return f"""전공자 수준 요약은 다음과 같은 기준을 따르라.\n
+                [학습자 정의]\n  
+                전공자는 문서에 포함된 개념, 구조, 논리 흐름을 고차적으로 분석하고, 이를 통합된 의미 체계로 재구성할 수 있는 사고 능력을 갖춘 학습자이다.\n 
+                [학습자 특성]\n  
+                -문서에 등장하는 중심 개념과 주변 개념들을 위계적으로 구조화하고, 인과 관계 또는 전개 순서를 파악할 수 있음.
+                -단락 간 논리 흐름, 저자의 전개 전략(비교-대조, 사례-추론 등)을 분석하고, 이 흐름에 따라 내용을 요약함.
+                -중심 개념이 문서 전체에서 어떤 역할을 수행하는지, 어떤 문제를 해결하려는지를 서술할 수 있음.
+                -문서에서 도출된 내용을 일반화하거나, 이론적·사회적·정책적 맥락으로 확장하여 재해석할 수 있음.
+                -단순 정보 나열이 아닌, 통찰과 구조를 중심으로 고차적 개념이 통합된 요약을 구성함.\n
+                [요약 수준 예시]\n
+                이 문서는 탄소중립을 단순한 온실가스 감축이 아니라, 에너지 체계·정책 구조·사회적 합의 메커니즘 전반의 전환을 요구하는 총체적 전략으로 제시한다. 
+                특히 '에너지 정의(energy justice)'를 중심 축으로 삼아, 재생에너지 전환·시민 참여·탄소 가격제 등의 요소를 상호 연결하여 통합적 접근을 보여준다. 
+                이러한 구조는 탄소중립을 단순한 환경 규제 차원이 아닌 사회 구조 개편의 프레임으로 재위치시키며, 기후 대응의 실효성과 정의 구현 간의 긴장 관계를 통찰한다."""
+
+## question_level 수준 정의
+def user_question_level(question_level):
+    if question_level == "비전공자":
+        return(
+            "비전공자는 다음과 같은 사고 구조를 가진 학습자이다.\n"
+            "-SOLO 1단계 (전구조 단계): 준비는 되어 있으나 주어진 과제를 적절히 다루지 못함.\n"
+            "예: 탄소중립과 관련 없는 것은?\n"
+            "-SOLO 2단계 (단일구조 단계): 주어진 과제에서 한 가지 특색이나 측면을 지적할 수 있으나, 사실이나 정보 간의 관련성을 파악하지 못함.\n"
+            "예: 탄소중립의 정의는?\n"
+            "-SOLO 3단계 (다중구조 단계): 주어진 과제에서 여러 가지 특색이나 차원을 지적할 수 있으나, 상호 연관시키지는 못함.\n"
+            "예: 탄소중립 실현 기술 중 적절한 것을 모두 고르시오.\n"
+        )
+    elif question_level == "전공자":
+            return (
+            "전공자는 다음과 같은 사고 구조를 가진 학습자이다.\n"
+            "-SOLO 4단계 (관계화 단계): 주어진 과제에서 몇 가지 특색이나 측면을 연관시켜 전체적인 구조를 인식하며 의미를 갖춤.\n"
+            "예: 탄소세가 탄소중립에 기여하는 이유는?\n"
+            "-SOLO 5단계 (일반화확장 단계):  주어진 과제에서 전체 구조를 더 수준 높은 추상적 수준으로 확장시켜 일반화함.\n"
+            " 예: 다른 국가에 탄소중립 전략을 일반화할 때 고려할 요소는?"
+            "-SOLO 6단계 (고차 창출): 주어진 과제에서 기존 개념을 비판적으로 검토하고, 새로운 대안을 생성함.\n"
+            "예: 기존 개념의 한계를 비판하고 새로운 대안을 구성하는 전략은?\n"
+            )
+        
+## field 특징 정의
+def field_features(field: str) -> str:
+    field_map = {
+        "언어": "해당 문서의 핵심 내용을 바탕으로, 사용된 전문 용어와 개념의 구성 방식을 언어적 구조와 의미의 관점에서 해석하라.",
+        "인문학": "해당 문서의 핵심 내용을 문화적 맥락에서 해석하며, 인간과 자연의 관계, 핵심 개념이 가진 문화적, 사유적 함의를 중심으로 해석하라.",
+        "사회": "해당 문서의 핵심 내용의 사회적 영향력을 바탕으로, 사회 문제 해결 가능성, 제도적 수용성, 공공성과의 관계를 중심으로 해석하라. ",
+        "경제": "해당 문서의 핵심 내용을 산업적 활용과 비용 구조 측면에서 분석하고, 시장에서의 가치 창출 가능성과 경제적 파급력을 중심으로 해석하라.",
+        "과학": "해당 문서의 핵심 내용이 과학적 이론 및 실험 설계와 어떻게 연결되는지를 바탕으로, 이론 검증의 논리성과 과학적 타당성을 중심으로 해석하라.",
+        "종교": "해당 문서의 핵심 내용이 자연 질서, 생명의 가치, 윤리적 책임성과 어떻게 연결되는지를 신학적 관점에서 해석하라.",
+        "철학": "해당 문서의 핵심 내용을 바탕으로, 존재로적, 인식론적, 기술철학적 관점에서 그 의미와 본질을 성찰적으로 해석하라.",
+        "공학": "해당 문서의 기술 구조와 설계 요소를 바탕으로, 기능적 효율성과 실현 가능성 및 완성도와 구현 방식을 중심으로 해석하라."
+    }
+    return field_map.get(field, "")
