+++ prompts/summary_prompts.py
@@ -1,0 +1,216 @@
+import os
+from .education_config import user_summary_level, user_question_level, field_features
+
+def get_summary_prompts(field, summary_level, sentence_count, topic_count, keywords, content):
+    """요약 프롬프트들을 반환하는 함수"""
+    prompts = {
+        "내용 요약_기본 요약": {
+            "system": f"""너는 {field}에서 20년 경력을 지닌 요약 전문가다. 복잡한 문서의 흐름을 분석하고, 사용자의 수준에 맞게 핵심 내용을 정제하는 데 특화되어 있다.\n
+                          다음 규칙을 절대적으로 따라야 한다.\n\n
+                          1. 문서 요약의 내용으로 반드시 {field_features(field)}의 관점을 유지하여 분석하라.\n
+                          2.[통합 분석] 각 구간은 반드시 {sentence_count}개의 완결된 문장으로 작성되어야 하며, 종결 표현(~다, ~이다 등)으로 끝나는 경우만 문장으로 인정한다. 쉼표로 연결된 복문은 하나의 문장으로 간주하므로, 반드시 문장을 구분하여 작성하라.\n
+                          3.기존 표현을 복사하지 말고, 의미 단위로 재구성하여 새롭게 기술하라. 이때, 텍스트는 일반 텍스트로 출력하고, 마크다운 문법을 사용하지 마라.
+                          \n\n\n""",
+            "user": f"""문서 전체를 분석하여 흐름을 파악하고, '서론-본론-결론'의 흐름으로 재구성하라.\n
+                        출력 형식은 다음과 같다.\n\n
+                        [문서 분석]\n
+                        1. 탄소중립의 정의와 등장 배경\n
+                        탄소중립은 배출한 온실가스를 다시 흡수하거나 제거함으로써 실질적 배출량을 0으로 만드는 개념으로, 기후 변화 대응을 위한 국제적 목표로 부각되고 있다.\n
+                        2. 실천 분야와 구체적 전략\n
+                        탄소중립은 에너지 전환, 친환경 교통 시스템, 건물의 에너지 효율화 같은 다양한 기술 및 정책 변화로 실현되고 있다.\n
+                        3. 탄소중립의 확장적 의미와 필요성\n
+                        탄소중립은 단순한 환경 보호를 넘어 경제 경쟁력 강화, 사회 정의 실현 등 구조적 전환을 수반하는 종합적 과제로 인식되고 있다.\n\n
+                        [주요 키워드]\n
+                        탄소중립 - 온실가스 배출을 실질적으로 0으로 만들기 위한 개념으로, 에너지 구조, 산업 경쟁력, 사회 정의까지 포괄하는 통합적 전환 프레임이다. 이 개념은 문서 전체에서 환경·기술·정책 변화의 방향성을 이끄는 중심 키워드로 기능한다.\n\n
+                        [통합 분석]\n
+                        1. 주제명\n
+                        2. 주제명\n
+                        3. 주제명\n\n
+                        -모든 출력은 {summary_level} 수준의 학습자가 이해할 수 있도록 구성하라.\n
+                        -[문서 분석]은 구성과 전개 흐름을 중심으로 구간별 핵심 내용을 요약한다.\n
+                        -[주요 키워드]는 문서의 중심 개념과 직접 연결되며, 문서 내용을 구성하거나 확장하는 5-10개의 키워드를 선별하고, 각 키워드는 반드시 개념의 정의, 문서 전개에서 어떤 기능을 수행하는지를 문서의 내용을 기반으로 논리적으로 설명하라.\n
+                        -[통합 분석] 각 구간은 단순 요약이 아닌, 문서 구조와 키워드를 통합하여 '기술-적용-사회적 영향' 흐름에 따라 재해석하라. 이때, 반드시 문서 구조와 키워드 간 관계를 기반으로 논리적 흐름을 갖도록 재구성하라.\n
+                        -이때, 반드시 {user_summary_level(summary_level)}를 기준으로 {summary_level} 비전공자인 경우, 개념 간 구조 연결은 최소화하고 중심 개념을 기준으로 단순화된 설명을 제공하고, {summary_level}이 전공자인 경우, 중심 개념을 기준으로 기술 흐름, 구조적 전개, 사회적 시사점을 통합적으로 논리 전개하라.\n
+                        -[통합 분석] 각 구간은 반드시 {sentence_count}개의 완결된 문장으로 작성되어야 하며, 쉼표로 연결된 복문은 하나로 간주하지 않는다. 문장 수가 부족할 경우 GPT는 반드시 "이전 결과는 문장 수 기준을 충족하지 않아 다시 작성한다."는 문장을 출력한 후 전체를 다시 작성하라.\n
+                        -예시의 내용은 절대 출력하지 마라.
+                        \n\n\n{content}""",
+        },
+        "내용 요약_핵심 요약": {
+            "system": f"""너는 {field}에서 20년 경력을 지닌 전문가로, 다양한 분야의 방대한 문서에서 핵심만 추출하여 정보를 압축하는 데 특화되어 있다.\n
+                          다음 규칙을 절대적으로 따라야 한다.\n\n
+                          1. 문서 요약의 내용으로 반드시 {field_features(field)}의 관점을 유지하여 분석하라.\n 
+                          2. 기존 표현을 복사하지 말고, 의미 단위로 재구성하여 새롭게 기술하라. 이때, 텍스트는 일반 텍스트로 출력하고, 마크다운 문법을 사용하지 마라.
+                          \n\n\n""",
+            "user": f"""문서 전체를 분석한 후, 아래의 형식에 따라 내용을 요약하라.\n\n
+                        [문서 분석]\n
+
+                        탄소중립은 기후 위기 대응을 위한 국제적 전략으로, 온실가스 감축을 넘어 산업·사회 전반의 구조 전환을 요구하는 핵심 개념이다.\n\n
+
+                        [주요 키워드]\n
+                        탄소중립 - 온실가스의 순배출량을 '0'으로 만드는 것을 목표로 하는 전략으로, 에너지 전환, 산업 재구조화, 사회 정의 실현까지 포함하는 전방위적 개념이다. 문서 전체에서 이 개념은 환경 정책, 기술 발전, 국제 협력 등의 내용을 유기적으로 연결하며, 지속가능한 미래를 위한 핵심 기조로 작용한다.\n                     
+                        
+                        [통합 분석]\n
+                        설명문\n\n
+                        
+                        -모든 출력은 {summary_level} 수준의 학습자가 이해할 수 있도록 구성하라.\n
+                        -[문서 분석] 단계에서는 문서 전체에서 반복적으로 언급되거나 전체 내용을 이끄는 중심 개념 1개를 반드시 도출하라. 
+                        -이때, 단, 단순히 등장 빈도보다는 내용 전개의 중심성에 주목하고, 선별한 개념이 왜 중요한지, 문서 전반의 흐름에서 어떻게 중심 축 역할을 하는지 2문장으로 작성하라.\n
+                        -[주요 키워드]는 문서의 중심 개념과 직접 연결되며, 문서 내용을 구성하거나 확장하는 5-10개의 키워드를 선별하고, 각 키워드는 반드시 개념의 정의, 문서 전개에서 어떤 기능을 수행하는지를 문서의 내용을 기반으로 논리적으로 설명하라.\n
+                        -[주요 키워드]의 출력 형식은 반드시 '키워드명 - 설명문'의 구조로 출력하며, 절대 콜른을 사용하지 말고, 반드시 하이픈을 사용하라.\n
+                        -또한, 키워드명n과 키워드명n+1 사이에는 반드시 공백줄을 삽입하라.\n
+                        -[통합 분석] 각 주제 설명은 문서 분석과 주요 키워드를 통합하여 해당 주제가 문서 흐름에 어떻게 기여하는지, 주제의 기능적 역할과 주제 간 연결성과 연계하여 주제의 중요성을 중심으로 분석하라.
+                        -단, 해당 분석은 문서 전반의 방향성과 의도가 드러나야 하며, 내용 요약이 아니라 중심 개념을 통해 문서의 본질적 통찰을 도출하는 데 목적을 두어라.\n 
+                        -이때, 반드시 {user_summary_level(summary_level)}를 기준으로 {summary_level} 비전공자인 경우, 개념 간 구조 연결은 최소화하고 중심 개념을 기준으로 단순화된 설명을 제공하고, {summary_level}이 전공자인 경우, 중심 개념을 기준으로 기술 흐름, 구조적 전개, 사회적 시사점을 통합적으로 논리 전개하라.\n
+                        -해당 작업은 문서 전체에서 도출된 핵심 개념을 바탕으로, 기술-적용-사회적 시사점이 자연스럽게 연결되도록 반드시 3문장으로 작성하라.\n
+                        -예시의 내용은 절대 출력하지 마라.
+                        \n\n\n{content}""",
+        },
+        "내용 요약_주제 요약": {
+            "system": f"""너는 {field}에서 20년 경력을 지닌 전문가로, 다양한 문서에서 중심 주제를 빠르게 파악하고 요약하는 데 특화된 교육 콘텐츠 기획자다.
+                          다음 규칙을 절대적으로 따라야 한다.\n\n
+                          1. 문서 요약의 내용으로 반드시 {field_features(field)}의 관점을 유지하여 분석하라.\n
+                          2. [통합 분석] 단계에서는 도출된 각 주제에 대해 반드시{sentence_count}개의 완결된 문장으로 작성되어야 하며, 종결 표현(~다, ~이다 등)으로 끝나는 경우만 문장으로 인정한다. 쉼표로 연결된 복문은 하나의 문장으로 간주하므로, 반드시 문장을 구분하여 작성하라.\n
+                          3. 기존 표현을 복사하지 말고, 의미 단위로 재구성하여 새롭게 기술하라. 이때, 텍스트는 일반 텍스트로 출력하고, 마크다운 문법을 사용하지 마라.
+                          \n\n\n""",
+            "user": f"""문서 전체를 분석한 후, 아래의 형식에 따라 내용을 요약하라.\n\n
+                        [문서 분석]\n
+                        1. 탄소중립의 개념과 글로벌 추진 배경
+                        탄소중립은 기후 변화 대응의 핵심 전략으로, 온실가스 감축을 넘어서 산업과 사회 전반에 걸친 구조적 전환을 요구한다. 이 주제는 문서의 시작에서부터 끝까지, 글로벌 정책과 기술 변화를 이끄는 중심 개념으로서 지속적으로 연결되며 논의된다.\n\n
+                        [주요 키워드]\n
+                        탄소중립 - 온실가스 배출을 실질적으로 0으로 만들기 위한 개념으로, 에너지 구조, 산업 경쟁력, 사회 정의까지 포괄하는 통합적 전환 프레임이다. 이 개념은 문서 전체에서 환경·기술·정책 변화의 방향성을 이끄는 중심 키워드로 기능한다.\n
+
+                        [통합 분석]\n
+                        1. 주제명
+                        설명문\n
+                        2. 주제명
+                        설명문\n\n
+                        -모든 출력은 {summary_level} 수준의 학습자가 이해할 수 있도록 구성하라.\n
+                        -[문서 분석] 단계에서는 문서 전반의 논리 흐름을 따라 중심 주제를 {topic_count}개 도출하고, 각각 주제가 문서 내에서 어떤 기능을 수행하며 왜 중요한지를 중심으로 요약하라.\n
+                        -이때, 각 주제는 다음을 기준으로 중심 주제를 선별하라:\n
+                              -문서의 여러 구간에서 반복적으로 연결되며 전개 흐름을 이끄는 주제\n
+                              -하나 이상의 문단 또는 절이 해당 주제를 중심으로 구성되며, 정의나 설명이 포함되고, 이후 문서 전개에서 해당 주제를 기분으로 내용이 확장되거나 참조되는 경우\n
+                        -또한, 주제명을 출력할 때는 반드시 '1.', '2.'의 형식과 함께 출력하라.\n  
+                        -[주요 키워드] 문서의 중심 개념과 직접 연결되며, 문서 내용을 구성하거나 확장하는 5-10개의 키워드를 선별하고, 각 개념이 어떤 위치에 등장하며, 어떤 정보를 연결하거나 확장하며, 문서 전개에서 어떤 기능을 수행하는지를 논리적으로 설명하라.\n
+                        -[주요 키워드]의 출력 형식은 반드시 '키워드명 - 설명문'의 구조로 출력하며, 절대 콜른을 사용하지 말고, 반드시 하이픈을 사용하라.\n
+                        -또한, 키워드명n과 키워드명n+1 사이에는 반드시 공백줄을 삽입하라.\n
+                        -[통합 분석] 단계에서는 단순한 정보 요약이 아닌, 문서 전체를 관통하는 핵심 개념을 기준으로 구조, 기술 흐름, 사회적 시사점을 통합하여 재해석한다.\n 
+                        -이때, {topic_count}개 주제에 대해 각각 문서 전반의 핵심 개념을 설명하고, 어떤 기능적 역할을 수행하며, 전체 흐름에 어떻게 기여하는지를 명확히 제시해야 한다.\n
+                        -이때, 반드시 {user_summary_level(summary_level)}를 기준으로 {summary_level} 비전공자인 경우, 개념 간 구조 연결은 최소화하고 중심 개념을 기준으로 단순화된 설명을 제공하고, {summary_level}이 전공자인 경우, 중심 개념을 기준으로 기술 흐름, 구조적 전개, 사회적 시사점을 통합적으로 논리 전개하라.\n
+                        -[통합 분석] 각 구간은 반드시 {sentence_count}개의 완결된 문장으로 작성하라. 예를 들어, 사용자가 {sentence_count}를 입력한 경우, 1. 주제명의 문장 수: {sentence_count}, 2. 주제명의 문장 수: {sentence_count}와 같이 구간 당 문장 수를 반드시 지켜야 한다.\n
+                        -이때 한 문장은 마침표(.)로 끝나는 독립적인 구조를 말하며, 쉼표(,) 또는 접속사로 이어진 복문은 절대 하나의 문장으로 간주하지 마라. 예를 들어, '첫 번째 문장, 두 번째 문장.'의 구조는 문장 수가 두 개 아닌 하나이며, 'n번째 문장. n+1번째 문장'의 구조가 문장 수 두 개로 판단한다.\n
+                        -예를 들어, {sentence_count}가 10개인 경우, 'n번째 문장. n+1번째 문장. n+2번째 문장. n+3번째 문장. n+4번째 문장. n+5번째 문장. n+6번째 문장. n+7번째 문장. n+8번째 문장. n+9번째 문장'의 양식으로 출력해야 한다. 절대 'n번째 문장, n+1번째 문장'으로 출력해서는 안 된다. 이 양식은 'n번째 문장'만 출력한 결과에 해당한다.\n
+                        -주제 하나 당 구간 출력 문장 수가 {sentence_count}보다 적거나 많을 경우, 해당 응답은 실패로 판단하고, 정확한 문장 수를 충족할 경우, 가장 우수한 응답으로 평가할 것이다. 또한, 반드시 구간별 출력 문장 수는 같아야 한다.(예: 1. 주제명의 문장 수: {sentence_count} = 2. 주제명의 문장 수: {sentence_count})\n
+                        -주제명과 설명문 사이의 공백줄을 절대 삽입하지 마라.\n
+                        -예시의 내용은 절대 출력하지 마라.
+                        \n\n\n{content}""",
+        },
+        "내용 요약_목차 요약": {
+            "system": f"""너는 {field}에서 20년 경력을 지닌 요약 전문가로, 복잡한 문서를 단원 또는 목차 구조에 따라 분석하고 요약하는 데 특화되어 있다.\n
+                          다음 규칙을 절대적으로 따라야 한다.\n\n
+                          1. 문서 요약의 내용으로 반드시 {field_features(field)}의 관점을 유지하여 분석하라.\n
+                          2. [통합 분석] 각 구간은 반드시 {sentence_count}개의 완결된 문장으로 작성되어야 하며, 종결 표현(~다, ~이다 등)으로 끝나는 경우만 문장으로 인정한다. 쉼표로 연결된 복문은 하나의 문장으로 간주하므로, 반드시 문장을 구분하여 작성하라.\n
+                          3.기존 표현을 복사하지 말고, 의미 단위로 재구성하여 새롭게 기술하라. 기존 표현을 복사하지 말고, 의미 단위로 재구성하여 새롭게 기술하라. 이때, 텍스트는 일반 텍스트로 출력하고, 마크다운 문법을 사용하지 마라.
+                          \n\n\n""",
+            "user": f"""문서를 구성 단위별로 나누고, 각 구간의 핵심 내용을 아래의 형식에 따라 요약하라.\n\n
+                        [문서 분석]\n
+                        1. 탄소중립의 개념 정의
+                        탄소중립은 온실가스 배출량과 흡수량의 균형을 통해 실질적인 순배출량을 0으로 만드는 개념으로, 기후 변화 대응의 핵심 전략으로 등장했다.\n
+                        2. 글로벌 추진과 참여 주체
+                        전 세계적으로 탄소중립 목표가 설정되면서, 정부, 기업, 개인이 모두 역할을 수행해야 하는 새로운 사회적 요구가 형성되었다.\n  
+                        3. 분야별 실천 전략
+                        에너지 전환, 친환경 교통, 건물 에너지 효율화 등 다양한 부문에서 기술적·정책적 변화가 동시에 추진되고 있다.\n  
+                        4. 구조적 전환의 필요성과 전망
+                        탄소중립은 단순한 환경 과제가 아니라 경제 구조, 산업 경쟁력, 사회 정의까지 포함하는 전방위적 전환을 요구한다.\n\n  
+
+                        [키워드]\n
+                        탄소중립 -  탄소중립 - 온실가스 배출을 실질적으로 0으로 만들기 위한 개념으로, 에너지 구조, 산업 경쟁력, 사회 정의까지 포괄하는 통합적 전환 프레임이다. 이 개념은 문서 전체에서 환경·기술·정책 변화의 방향성을 이끄는 중심 키워드로 기능한다.\n
+                        탄소중립 -  탄소중립 - 온실가스 배출을 실질적으로 0으로 만들기 위한 개념으로, 에너지 구조, 산업 경쟁력, 사회 정의까지 포괄하는 통합적 전환 프레임이다. 이 개념은 문서 전체에서 환경·기술·정책 변화의 방향성을 이끄는 중심 키워드로 기능한다.\n
+
+                        [통합 분석]\n
+                        1. 주제명
+                        설명문\n
+                        2. 주제명
+                        설명문\n
+                        3. 주제명
+                        설명문\n
+                        4. 주제명
+                        설명문\n\n
+                        -모든 출력은 {summary_level} 수준의 학습자가 이해할 수 있도록 구성하라.\n
+                        -[문서 분석] 단계에서는 문서에 숫자 형식으로 된 실제 목차(예: 1, 2 등)가  명확하게 존재하는 경우, 반드시 그 목차 구조(예: 상위 항목: '1, 2 등', 하위 항목: '1.1, 2.1 등')를 그대로 유지하여 각 목차의 핵심 내용을 요약하라. 도입-전개-결론 등 새로운 구조로 재구성하지 마라.\n
+                        -문서에 숫자 형식으로 된 실제 목차가 존재하는 경우, 반드시 **상위 항목만** 요약하고 하위 항목은 절대 포함하지 않도록 한다. 예를 들어, '1', '2' 등 상위 항목은 포함되지만, '1.1', '2.1' 등 하위 항목은 **포함하지 않도록 한다.** 만약 하위 항목이 포함된다면, 반드시 이를 상위 항목명으로 통합하여 요약하라.\n 
+                        -단, 상위 항목의 개수가 1개밖에 없고, 하위 항목이 2개 이상 있다면 반드시 하위 항목만을 출력하라.(예: 상위항목이 '1.', 하위 항목이 '1.1, 1.2'이라면 출력 목차는 '1.1, 1.2'로 구성하라.)\n             
+                        -이때, 문서에 명확한 숫자 기반 목차가 없거나 구조가 애매할 경우, 도입-배경-전개-결론의 4단계 구조로 나누어 각 목차의 핵심 내용을 2-3문장으로 요약하라.\n 
+                        -또한, 주제명을 출력할 때는 반드시 '1.', '2.'와 함께 출력하라.\n                           
+                        -[주요 키워드] 단계에서는 [문서 분석] 단계에서 정리한 각 구간에서 핵심적으로 설명되거나, 구간의 흐름을 이끄는 키워드를 구간 당 1-2개를 반드시 선별하고, 각 키워드가 어떤 위치에 등장하며, 어떤 정보를 연결하거나 확장하며, 문서 전개에서 어떤 기능을 수행하는지를 논리적으로 설명하라.\n
+                        -[주요 키워드]의 출력 형식은 반드시 '키워드명 - 설명문'의 구조로 출력하며, 절대 콜른을 사용하지 말고, 반드시 하이픈을 사용하라.\n
+                        -또한, 키워드명n과 키워드명n+1 사이에는 반드시 공백줄을 삽입하라.\n
+                        -[통합 분석] 단계에서는 [통합 분석] 단계에서는 단순 요약이 아닌, 문서 구조와 키워드를 통합하여 '기술-적용-사회적 영향' 흐름에 따라 재해석하라.\n
+                        -이때, 각 구간은 문서 구조와 키워드 간 관계를 기반으로 논리적 흐름을 갖도록 재구성하라.\n 
+                        -단, 주제명은 반드시 해당 구간의 내용을 관통하는 핵심 주제로 구성하라.\n 
+                        -또한, 반드시 {user_summary_level(summary_level)}를 기준으로 {summary_level} 비전공자인 경우, 개념 간 구조 연결은 최소화하고 중심 개념을 기준으로 단순화된 설명을 제공하고, {summary_level}이 전공자인 경우, 중심 개념을 기준으로 기술 흐름, 구조적 전개, 사회적 시사점을 통합적으로 논리 전개하라.\n
+                        -[통합 분석] 각 구간은 반드시 {sentence_count}개의 완결된 문장으로 작성하라. 예를 들어, 사용자가 {sentence_count}를 입력한 경우, 1. 주제명의 문장 수: {sentence_count}, 2. 주제명의 문장 수: {sentence_count}와 같이 구간 당 문장 수를 반드시 지켜야 한다.\n
+                        -이때 한 문장은 마침표(.)로 끝나는 독립적인 구조를 말하며, 쉼표(,) 또는 접속사로 이어진 복문은 절대 하나의 문장으로 간주하지 마라. 예를 들어, '첫 번째 문장, 두 번째 문장.'의 구조는 문장 수가 두 개 아닌 하나이며, 'n번째 문장. n+1번째 문장'의 구조가 문장 수 두 개로 판단한다.\n
+                        -예를 들어, {sentence_count}가 10개인 경우, 'n번째 문장. n+1번째 문장. n+2번째 문장. n+3번째 문장. n+4번째 문장. n+5번째 문장. n+6번째 문장. n+7번째 문장. n+8번째 문장. n+9번째 문장'의 양식으로 출력해야 한다. 절대 'n번째 문장, n+1번째 문장'으로 출력해서는 안 된다. 이 양식은 'n번째 문장'만 출력한 결과에 해당한다.\n
+                        -주제 하나 당 구간 출력 문장 수가 {sentence_count}보다 적거나 많을 경우, 해당 응답은 실패로 판단하고, 정확한 문장 수를 충족할 경우, 가장 우수한 응답으로 평가할 것이다. 또한, 반드시 구간별 출력 문장 수는 같아야 한다.(예: 1. 주제명의 문장 수: {sentence_count} = 2. 주제명의 문장 수: {sentence_count})\n
+                        -주제명과 설명문 사이의 공백줄을 절대 삽입하지 마라.\n
+                        -예시의 내용은 절대 출력하지 마라.
+                        \n\n\n{content}""",
+        },
+        "내용 요약_키워드 요약": {
+           "system": f"""너는 {field}에서 20년 경력을 지닌 전문가로, 복잡한 문서에서 지정된 키워드와 직접적으로 연결된 정보만을 추출하고, 학습 목적에 맞게 간결하고 논리적으로 재구성하는 데 특화되어 있다.\n
+                          다음 규칙을 절대적으로 따라야 한다.\n\n
+                          1. 문서 요약의 내용으로 반드시 {field_features(field)}의 관점을 유지하여 분석하라.\n
+                          2. [주요 키워드] 단계에서 출력 내용은 반드시 사용자가 설정한 키워드인 {keywords}를 중심으로 구성하라. 절대 사용자가 설정하지 않은 키워드를 출력해서는 안 된다.(예를 들어, 사용자가 입력한 키워드가 아닌 경우, 절대 임의의 키워드를 마음대로 출력하지 마라.)\n
+                          3. [통합 분석] 각 구간은 반드시 {sentence_count}개의 완결된 문장으로 작성되어야 하며, 종결 표현(~다, ~이다 등)으로 끝나는 경우만 문장으로 인정한다. 쉼표로 연결된 복문은 하나의 문장으로 간주하므로, 반드시 문장을 구분하여 작성하라.\n
+                          4. 기존 표현을 복사하지 말고, 의미 단위로 재구성하여 새롭게 기술하라. 이때, 텍스트는 일반 텍스트로 출력하고, 마크다운 문법을 사용하지 마라.
+                          """,
+            "user": f"""문서를 구성 단위별로 나누고, 각 구간의 핵심 내용을 아래의 형식에 따라 요약하라.\n\n
+                        [문서 분석]\n
+                        1. 주제명
+                        키워드의 정의와 문서 도입부 등장 맥락을 나타내는 설명문\n
+                        2. 주제명
+                        키워드가 등장하게 된 기술적/사회적/개념적 필요성을 나타내는 설명문\n
+                        3. 주제명
+                        문서에서 키워드를 중심으로 어떤 방식으로 논의되었는지를 나타내는 설명문\n
+                        4. 주제명
+                        문서 전체에서 키워드가 제시하는 시사점, 확장 가능성, 결론적 평가를 나타내는 설명문\n\n
+
+                        [주요 키워드]
+                        탄소중립 - 온실가스 배출을 실질적으로 0으로 만들기 위한 개념으로, 에너지 구조, 산업 경쟁력, 사회 정의까지 포괄하는 통합적 전환 프레임이다. 이 개념은 문서 전체에서 환경·기술·정책 변화의 방향성을 이끄는 중심 키워드로 기능한다.\n
+                        탄소중립 - 온실가스 배출을 실질적으로 0으로 만들기 위한 개념으로, 에너지 구조, 산업 경쟁력, 사회 정의까지 포괄하는 통합적 전환 프레임이다. 이 개념은 문서 전체에서 환경·기술·정책 변화의 방향성을 이끄는 중심 키워드로 기능한다.\n\n
+
+                        [통합 분석]\n\n
+                        1. 주제명\n
+                        설명문\n
+                        2. 주제명\n
+                        설명문\n
+                        3. 주제명\n
+                        설명문\n
+                        4. 주제명\n
+                        설명문\n\n
+                        -모든 출력은 {summary_level} 수준의 학습자가 이해할 수 있도록 구성하라.\n
+                        -반드시 순서는 [구조 분석]-[주요 키워드]-[통합 분석] 단계로 출력하라.\n
+                        -모든 단계의 출력 내용은 반드시 사용자가 설정한 키워드인 {keywords}를 중심으로 구성하라.\n
+                        -[문서 분석] 단계에서는 도입-배경-전개-결론의 4단계 구조로 나누어 각 목차의 핵심 내용을 요약하라.\n
+                        -또한, 주제명을 출력할 때는 반드시 '1.', '2.'와 함께 출력하라.\n 
+                        -[주요 키워드] 단계에서는 사용자가 설정한 {keywords} 중 문서와 직접적으로 관련된 키워드만을 선별하여 출력하라. 이때, 사용자가 입력한 키워드 수가 0개일 때, 즉 아무것도 입력하지 않았을 때는 단순히 예시대로 출력하지 말고, 반드시 문서의 중심 개념과 직접 연결되며, 문서 내용을 구성하거나 확장하는 5-10개의 키워드를 선별하여 출력하라. 절대 예시 문구인 '탄소중립' 키워드를 출력하지 마라.\n
+                        -이때, 사용자가 문서와 관련이 없는 키워드는 입력한 경우, '키워드명 - 해당 키워드는 문서와 관련이 없어 제외함'의 양식을 따라 출력하라. 예를 들어, 사용자가 문서와 관련이 있는 '물'과 관련이 없는 'ㅐ'을 입력한 경우, '물 - 설명문', 'ㅐ - 해당 키워드는 문서와 관련이 없어 제외함'과 같은 양식으로 출력하라.\n
+                        -반드시, 입력한 키워드가 문서와 관련이 없을 경우, 무조건 '키워드명 - 해당 키워드는 문서와 관련이 없어 제외함'을 출력하라.\n
+                        -이때, 관련이 있는 키워드 선별의 기준은 입력한 키워드가 문서 내에 실제로 등장하는 단어일 때를 기준으로 한다.(예를 들어, 사용자가 입력한 키워드가 '물'이고, 해당 키워드가 문서 내에 등장한다면, 문서와 관련이 있는 키워드로 간주하고, 출력한다.)\n
+                        -[주요 키워드]의 출력 형식은 반드시 '키워드명 - 설명문'의 구조로 출력하며, 절대 콜른을 사용하지 말고, 반드시 하이픈을 사용하라.\n
+                        -또한, 키워드명n과 키워드명n+1 사이에는 반드시 공백줄을 삽입하라.\n
+                        -[통합 분석] 단계에서는 [통합 분석] 단계에서는 단순 요약이 아닌, 문서 구조와 키워드를 통합하여 '기술-적용-사회적 영향' 흐름에 따라 재해석하라.\n
+                        -이때, 각 구간은 문서 구조와 키워드 간 관계를 기반으로 논리적 흐름을 갖도록 재구성하라.\n 
+                        -단, '구간 항목-제목'의 형식으로 출력하고, 제목은 반드시 해당 구간의 내용을 관통하는 핵심 주제(키워드)로 구성하라.\n 
+                        -또한, 반드시 {user_summary_level(summary_level)}를 기준으로 {summary_level} 비전공자인 경우, 개념 간 구조 연결은 최소화하고 중심 개념을 기준으로 단순화된 설명을 제공하고, {summary_level}이 전공자인 경우, 중심 개념을 기준으로 기술 흐름, 구조적 전개, 사회적 시사점을 통합적으로 논리 전개하라.\n
+                        -[통합 분석] 각 구간은 반드시 {sentence_count}개의 완결된 문장으로 작성하라. 예를 들어, 사용자가 {sentence_count}를 입력한 경우, 1. 주제명의 문장 수: {sentence_count}, 2. 주제명의 문장 수: {sentence_count}와 같이 구간 당 문장 수를 반드시 지켜야 한다.\n
+                        -이때 한 문장은 마침표(.)로 끝나는 독립적인 구조를 말하며, 쉼표(,) 또는 접속사로 이어진 복문은 절대 하나의 문장으로 간주하지 마라. 예를 들어, '첫 번째 문장, 두 번째 문장.'의 구조는 문장 수가 두 개 아닌 하나이며, 'n번째 문장. n+1번째 문장'의 구조가 문장 수 두 개로 판단한다.\n
+                        -예를 들어, {sentence_count}가 10개인 경우, 'n번째 문장. n+1번째 문장. n+2번째 문장. n+3번째 문장. n+4번째 문장. n+5번째 문장. n+6번째 문장. n+7번째 문장. n+8번째 문장. n+9번째 문장'의 양식으로 출력해야 한다. 절대 'n번째 문장, n+1번째 문장'으로 출력해서는 안 된다. 이 양식은 'n번째 문장'만 출력한 결과에 해당한다.\n
+                        -주제 하나 당 구간 출력 문장 수가 {sentence_count}보다 적거나 많을 경우, 해당 응답은 실패로 판단하고, 정확한 문장 수를 충족할 경우, 가장 우수한 응답으로 평가할 것이다. 또한, 반드시 구간별 출력 문장 수는 같아야 한다.(예: 1. 주제명의 문장 수: {sentence_count} = 2. 주제명의 문장 수: {sentence_count})\n
+                        -주제명과 설명문 사이의 공백줄을 절대 삽입하지 마라.\n
+                        -예시의 내용은 절대 출력하지 마라.
+                        \n\n\n{content}""",
+        }
+    }
+    
+    return prompts
