+++ utils.py
@@ -1,708 +1,13 @@
-import os
-from openai import OpenAI
-import tiktoken
-import fitz
-from pptx import Presentation
-from settings import (
-    API_KEY, 
-    MODEL_NAME, 
-    MAX_USER_MESSAGE_TOKENS,
-    BASE_URL
-)
+from core.openai_client import get_base_url, truncate_by_tokens, summarize_with_chatgpt
+from core.file_processor import extract_text_from_pdf, extract_text_from_pptx
+from prompts.prompt_manager import get_summary_prompt
 
-# OpenAI 클라이언트 초기화
-client = OpenAI(api_key=API_KEY)
-
-# URL 관련 함수
-def get_base_url():
-    return BASE_URL
-
-# PDF에서 텍스트 추출
-def extract_text_from_pdf(file_path):
-    text = ""
-    with fitz.open(file_path) as doc:
-        for page in doc:
-            text += page.get_text()
-    return text
-
-# PPTX에서 텍스트 추출
-def extract_text_from_pptx(file_path):
-    text = ""
-    prs = Presentation(file_path)
-    for slide in prs.slides:
-        for shape in slide.shapes:
-            if hasattr(shape, "text"):
-                text += shape.text + "\n"
-    return text
-
-# 토큰 제한 함수
-def truncate_by_tokens(text, max_tokens, model=MODEL_NAME):
-    encoding = tiktoken.encoding_for_model(model)
-    tokens = encoding.encode(text)
-    if len(tokens) > max_tokens:
-        tokens = tokens[:max_tokens]
-    return encoding.decode(tokens)
-
-# GPT 요약 함수
-def summarize_with_chatgpt(system_message, user_message):
-    user_message = truncate_by_tokens(user_message, MAX_USER_MESSAGE_TOKENS)
-
-    completion = client.chat.completions.create(
-        model=MODEL_NAME,
-        messages=[
-            {"role": "system", "content": system_message},
-            {"role": "user", "content": user_message},
-            
-        ],
-        temperature=0.7,
-    )
-
-    summary = completion.choices[0].message.content
-    usage = completion.usage
-    return summary, usage
-
-# 요약 프롬프트 구성
-def get_summary_prompt(type_name, content):
-    # 환경 변수에서 설정값 가져오기
-    field = os.getenv("field", "공학")
-    summary_level = os.getenv("summary_level", "전공자")
-    question_level = os.getenv("question_level", "전공자")
-    sentence_count = os.getenv("sentence_count", "500")
-    topic_count = os.getenv("topic_count", "2")
-    keywords = os.getenv("keywords", "기타,등등")
-    question_count = os.getenv("question_count", "3")
-    choice_count = os.getenv("choice_count", "4")
-    choice_format = os.getenv("choice_format", "문장형")
-    array_choice_count = os.getenv("array_choice_count", "3")
-    blank_count = os.getenv("blank_count", "1")
-
-    if isinstance(keywords, str) and "," in keywords:
-        keywords = keywords.split(",")
-
-    ## summary_level 수준 정의
-    def user_summary_level(summary_level):
-        if summary_level == "비전공자":
-            return f""" 비전공자 수준 요약은 다음과 같은 기준을 따르라.\n
-                    [학습자 정의]\n
-                    비전공자는 문서에 제시된 정보나 용어를 사실적으로 이해할 수 있으나, 개념 간 관계를 구조화하거나 의미를 확장하는 데에는 어려움을 겪는 학습자이다.\n
-                    [학습자 특성]\n
-                    -문서에 명시된 정의, 사례, 수치, 설명 등을 표면적으로 요약할 수 있음.
-                    -서술된 정보는 항목별로 나열하되, 항목 간의 연결성이나 구조적 흐름은 형성하지 못함.
-                    -문서의 중심 개념을 반복 횟수에 따라 중요하게 인식하지만, 그 개념이 수행하는 기능이나 맥락적 의미는 분석하지 않음.
-                    -문서 전체의 의도, 전개 방식, 개념적 구조를 통합적으로 파악하지 못함.
-                    -전문 용어 사용을 지양하고, 직관적이고 쉬운 표현 위주로 요약을 구성함.\n
-                    [요약 수준 예시]\n
-                    탄소중립은 온실가스를 줄이거나 흡수해서 실제 배출량을 0으로 만드는 것이다. 
-                    이를 위해 교통, 에너지, 건물 등에서 다양한 정책이 시행되고 있다. 우리나라뿐 아니라 여러 나라가 기후변화를 막기 위해 노력하고 있다."""
-
-        elif summary_level == "전공자":
-            return f"""전공자 수준 요약은 다음과 같은 기준을 따르라.\n
-                    [학습자 정의]\n  
-                    전공자는 문서에 포함된 개념, 구조, 논리 흐름을 고차적으로 분석하고, 이를 통합된 의미 체계로 재구성할 수 있는 사고 능력을 갖춘 학습자이다.\n 
-                    [학습자 특성]\n  
-                    -문서에 등장하는 중심 개념과 주변 개념들을 위계적으로 구조화하고, 인과 관계 또는 전개 순서를 파악할 수 있음.
-                    -단락 간 논리 흐름, 저자의 전개 전략(비교-대조, 사례-추론 등)을 분석하고, 이 흐름에 따라 내용을 요약함.
-                    -중심 개념이 문서 전체에서 어떤 역할을 수행하는지, 어떤 문제를 해결하려는지를 서술할 수 있음.
-                    -문서에서 도출된 내용을 일반화하거나, 이론적·사회적·정책적 맥락으로 확장하여 재해석할 수 있음.
-                    -단순 정보 나열이 아닌, 통찰과 구조를 중심으로 고차적 개념이 통합된 요약을 구성함.\n
-                    [요약 수준 예시]\n
-                    이 문서는 탄소중립을 단순한 온실가스 감축이 아니라, 에너지 체계·정책 구조·사회적 합의 메커니즘 전반의 전환을 요구하는 총체적 전략으로 제시한다. 
-                    특히 '에너지 정의(energy justice)'를 중심 축으로 삼아, 재생에너지 전환·시민 참여·탄소 가격제 등의 요소를 상호 연결하여 통합적 접근을 보여준다. 
-                    이러한 구조는 탄소중립을 단순한 환경 규제 차원이 아닌 사회 구조 개편의 프레임으로 재위치시키며, 기후 대응의 실효성과 정의 구현 간의 긴장 관계를 통찰한다."""
-
-    ## question_level 수준 정의
-    def user_question_level(question_level):
-        if question_level == "비전공자":
-            return(
-                "비전공자는 다음과 같은 사고 구조를 가진 학습자이다.\n"
-                "-SOLO 1단계 (전구조 단계): 준비는 되어 있으나 주어진 과제를 적절히 다루지 못함.\n"
-                "예: 탄소중립과 관련 없는 것은?\n"
-                "-SOLO 2단계 (단일구조 단계): 주어진 과제에서 한 가지 특색이나 측면을 지적할 수 있으나, 사실이나 정보 간의 관련성을 파악하지 못함.\n"
-                "예: 탄소중립의 정의는?\n"
-                "-SOLO 3단계 (다중구조 단계): 주어진 과제에서 여러 가지 특색이나 차원을 지적할 수 있으나, 상호 연관시키지는 못함.\n"
-                "예: 탄소중립 실현 기술 중 적절한 것을 모두 고르시오.\n"
-            )
-        elif question_level == "전공자":
-                return (
-                "전공자는 다음과 같은 사고 구조를 가진 학습자이다.\n"
-                "-SOLO 4단계 (관계화 단계): 주어진 과제에서 몇 가지 특색이나 측면을 연관시켜 전체적인 구조를 인식하며 의미를 갖춤.\n"
-                "예: 탄소세가 탄소중립에 기여하는 이유는?\n"
-                "-SOLO 5단계 (일반화확장 단계):  주어진 과제에서 전체 구조를 더 수준 높은 추상적 수준으로 확장시켜 일반화함.\n"
-                " 예: 다른 국가에 탄소중립 전략을 일반화할 때 고려할 요소는?"
-                "-SOLO 6단계 (고차 창출): 주어진 과제에서 기존 개념을 비판적으로 검토하고, 새로운 대안을 생성함.\n"
-                "예: 기존 개념의 한계를 비판하고 새로운 대안을 구성하는 전략은?\n"
-                )
-        
-    ## field 특징 정의
-    def field_features(field: str) -> str:
-        field_map = {
-            "언어": "해당 문서의 핵심 내용을 바탕으로, 사용된 전문 용어와 개념의 구성 방식을 언어적 구조와 의미의 관점에서 해석하라.",
-            "인문학": "해당 문서의 핵심 내용을 문화적 맥락에서 해석하며, 인간과 자연의 관계, 핵심 개념이 가진 문화적, 사유적 함의를 중심으로 해석하라.",
-            "사회": "해당 문서의 핵심 내용의 사회적 영향력을 바탕으로, 사회 문제 해결 가능성, 제도적 수용성, 공공성과의 관계를 중심으로 해석하라. ",
-            "경제": "해당 문서의 핵심 내용을 산업적 활용과 비용 구조 측면에서 분석하고, 시장에서의 가치 창출 가능성과 경제적 파급력을 중심으로 해석하라.",
-            "과학": "해당 문서의 핵심 내용이 과학적 이론 및 실험 설계와 어떻게 연결되는지를 바탕으로, 이론 검증의 논리성과 과학적 타당성을 중심으로 해석하라.",
-            "종교": "해당 문서의 핵심 내용이 자연 질서, 생명의 가치, 윤리적 책임성과 어떻게 연결되는지를 신학적 관점에서 해석하라.",
-            "철학": "해당 문서의 핵심 내용을 바탕으로, 존재로적, 인식론적, 기술철학적 관점에서 그 의미와 본질을 성찰적으로 해석하라.",
-            "공학": "해당 문서의 기술 구조와 설계 요소를 바탕으로, 기능적 효율성과 실현 가능성 및 완성도와 구현 방식을 중심으로 해석하라."
-        }
-        return field_map.get(field, "")
-
-    prompts = {
-        "내용 요약_기본 요약": {
-             "system": f"""너는 {field}에서 20년 경력을 지닌 요약 전문가다. 복잡한 문서의 흐름을 분석하고, 사용자의 수준에 맞게 핵심 내용을 정제하는 데 특화되어 있다.
-                          다음 규칙을 절대적으로 따라야 한다.
-
-                          1. 문서 요약의 내용으로 반드시 {field_features(field)}의 관점을 유지하여 분석하라.
-                          2. [통합 분석] 각 구간은 반드시 {sentence_count}개의 완결된 문장으로 작성되어야 하며, 종결 표현(~다, ~이다 등)으로 끝나는 경우만 문장으로 인정한다. 쉼표로 연결된 복문은 하나의 문장으로 간주하므로, 반드시 문장을 구분하여 작성하라.
-                          3. 기존 표현을 복사하지 말고, 의미 단위로 재구성하여 새롭게 기술하라.
-                          4. 반드시 JSON 형태로 출력하라.
-                          
-                          출력 JSON 형식:
-                          {{
-                                "summary": {{
-                                    "summary_generation_type": "Content Summary_Basic Summary",
-                                    "summary_settings": {{
-                                        "domain": "{field}",
-                                        "summary_level": "{summary_level}",
-                                        "sentence_count": "{sentence_count}"
-                                    }},
-                                    "summary_content": {{
-                                        "document_analysis": [
-                                            {{"order": "1", "topic": "Topic1", "description": "Description1"}},
-                                            {{"order": "2", "topic": "Topic2", "description": "Description2"}},
-                                            {{"order": "3", "topic": "Topic3", "description": "Description3"}}
-                                        ],
-                                        "main_keywords": [
-                                            {{"keyword": "Keyword1", "description": "Keyword description1"}},
-                                            {{"keyword": "Keyword2", "description": "Keyword description2"}}
-                                        ],
-                                        "integrated_analysis": [
-                                            {{"order": "1", "topic": "Topic1", "analysis": "Analysis1"}},
-                                            {{"order": "2", "topic": "Topic2", "analysis": "Analysis2"}},
-                                            {{"order": "3", "topic": "Topic3", "analysis": "Analysis3"}}
-                                        ]
-                                    }}
-                                }}
-                          }}""",
-            "user": f"""문서 전체를 분석하여 흐름을 파악하고, '서론-본론-결론'의 흐름으로 재구성하라.
-                        모든 출력은 {summary_level} 수준의 학습자가 이해할 수 있도록 구성하라.
-                        {user_summary_level(summary_level)}를 기준으로 구성하라.
-                        통합 분석 각 구간은 반드시 {sentence_count}개의 완결된 문장으로 작성하라.
-                        \n\n\n{content}""",
-        },
-        "내용 요약_핵심 요약": {
-            "system": f"""너는 {field}에서 20년 경력을 지닌 전문가로, 다양한 분야의 방대한 문서에서 핵심만 추출하여 정보를 압축하는 데 특화되어 있다.\n
-                          다음 규칙을 절대적으로 따라야 한다.\n\n
-                          1. 문서 요약의 내용으로 반드시 {field_features(field)}의 관점을 유지하여 분석하라.\n 
-                          2. 기존 표현을 복사하지 말고, 의미 단위로 재구성하여 새롭게 기술하라. 이때, 텍스트는 일반 텍스트로 출력하고, 마크다운 문법을 사용하지 마라.
-                          \n\n\n""",
-            "user": f"""문서 전체를 분석한 후, 아래의 형식에 따라 내용을 요약하라.\n\n
-                        [문서 분석]\n
-
-                        탄소중립은 기후 위기 대응을 위한 국제적 전략으로, 온실가스 감축을 넘어 산업·사회 전반의 구조 전환을 요구하는 핵심 개념이다.\n\n
-
-                        [주요 키워드]\n
-                        탄소중립 - 온실가스의 순배출량을 '0'으로 만드는 것을 목표로 하는 전략으로, 에너지 전환, 산업 재구조화, 사회 정의 실현까지 포함하는 전방위적 개념이다. 문서 전체에서 이 개념은 환경 정책, 기술 발전, 국제 협력 등의 내용을 유기적으로 연결하며, 지속가능한 미래를 위한 핵심 기조로 작용한다.\n
-                        탄소중립 - 온실가스의 순배출량을 '0'으로 만드는 것을 목표로 하는 전략으로, 에너지 전환, 산업 재구조화, 사회 정의 실현까지 포함하는 전방위적 개념이다. 문서 전체에서 이 개념은 환경 정책, 기술 발전, 국제 협력 등의 내용을 유기적으로 연결하며, 지속가능한 미래를 위한 핵심 기조로 작용한다.\n                        
-                        
-                        [통합 분석]\n
-                        설명문\n\n
-                        
-                        -모든 출력은 {summary_level} 수준의 학습자가 이해할 수 있도록 구성하라.\n
-                        -[문서 분석] 단계에서는 문서 전체에서 반복적으로 언급되거나 전체 내용을 이끄는 중심 개념 1개를 반드시 도출하라. 
-                        -이때, 단, 단순히 등장 빈도보다는 내용 전개의 중심성에 주목하고, 선별한 개념이 왜 중요한지, 문서 전반의 흐름에서 어떻게 중심 축 역할을 하는지 2문장으로 작성하라.\n
-                        -[주요 키워드]는 문서의 중심 개념과 직접 연결되며, 문서 내용을 구성하거나 확장하는 5-10개의 키워드를 선별하고, 각 키워드는 반드시 개념의 정의, 문서 전개에서 어떤 기능을 수행하는지를 문서의 내용을 기반으로 논리적으로 설명하라.\n
-                        -[주요 키워드]의 출력 형식은 반드시 '키워드명 - 설명문'의 구조로 출력하며, 절대 콜른을 사용하지 말고, 반드시 하이픈을 사용하라.\n
-                        -또한, 키워드명n과 키워드명n+1 사이에는 반드시 공백줄을 삽입하라.\n
-                        -[통합 분석] 각 주제 설명은 문서 분석과 주요 키워드를 통합하여 해당 주제가 문서 흐름에 어떻게 기여하는지, 주제의 기능적 역할과 주제 간 연결성과 연계하여 주제의 중요성을 중심으로 분석하라.
-                        -단, 해당 분석은 문서 전반의 방향성과 의도가 드러나야 하며, 내용 요약이 아니라 중심 개념을 통해 문서의 본질적 통찰을 도출하는 데 목적을 두어라.\n 
-                        -이때, 반드시 {user_summary_level(summary_level)}를 기준으로 {summary_level} 비전공자인 경우, 개념 간 구조 연결은 최소화하고 중심 개념을 기준으로 단순화된 설명을 제공하고, {summary_level}이 전공자인 경우, 중심 개념을 기준으로 기술 흐름, 구조적 전개, 사회적 시사점을 통합적으로 논리 전개하라.\n
-                        -해당 작업은 문서 전체에서 도출된 핵심 개념을 바탕으로, 기술-적용-사회적 시사점이 자연스럽게 연결되도록 반드시 3문장으로 작성하라.\n
-                        -예시의 내용은 절대 출력하지 마라.
-                        \n\n\n{content}""",
-        },
-        "내용 요약_주제 요약": {
-            "system": f"""너는 {field}에서 20년 경력을 지닌 전문가로, 다양한 문서에서 중심 주제를 빠르게 파악하고 요약하는 데 특화된 교육 콘텐츠 기획자다.
-                          다음 규칙을 절대적으로 따라야 한다.\n\n
-                          1. 문서 요약의 내용으로 반드시 {field_features(field)}의 관점을 유지하여 분석하라.\n
-                          2. [통합 분석] 단계에서는 도출된 각 주제에 대해 반드시{sentence_count}개의 완결된 문장으로 작성되어야 하며, 종결 표현(~다, ~이다 등)으로 끝나는 경우만 문장으로 인정한다. 쉼표로 연결된 복문은 하나의 문장으로 간주하므로, 반드시 문장을 구분하여 작성하라.\n
-                          3. 기존 표현을 복사하지 말고, 의미 단위로 재구성하여 새롭게 기술하라. 이때, 텍스트는 일반 텍스트로 출력하고, 마크다운 문법을 사용하지 마라.
-                          \n\n\n""",
-            "user": f"""문서 전체를 분석한 후, 아래의 형식에 따라 내용을 요약하라.\n\n
-                        [문서 분석]\n
-                        1. 탄소중립의 개념과 글로벌 추진 배경
-                        탄소중립은 기후 변화 대응의 핵심 전략으로, 온실가스 감축을 넘어서 산업과 사회 전반에 걸친 구조적 전환을 요구한다. 이 주제는 문서의 시작에서부터 끝까지, 글로벌 정책과 기술 변화를 이끄는 중심 개념으로서 지속적으로 연결되며 논의된다.\n\n
-                        [주요 키워드]\n
-                        탄소중립 - 온실가스 배출을 실질적으로 0으로 만들기 위한 개념으로, 에너지 구조, 산업 경쟁력, 사회 정의까지 포괄하는 통합적 전환 프레임이다. 이 개념은 문서 전체에서 환경·기술·정책 변화의 방향성을 이끄는 중심 키워드로 기능한다.\n
-                        탄소중립 - 온실가스 배출을 실질적으로 0으로 만들기 위한 개념으로, 에너지 구조, 산업 경쟁력, 사회 정의까지 포괄하는 통합적 전환 프레임이다. 이 개념은 문서 전체에서 환경·기술·정책 변화의 방향성을 이끄는 중심 키워드로 기능한다.\n
-
-                        [통합 분석]\n
-                        1. 주제명
-                        설명문\n
-                        2. 주제명
-                        설명문\n\n
-                        -모든 출력은 {summary_level} 수준의 학습자가 이해할 수 있도록 구성하라.\n
-                        -[문서 분석] 단계에서는 문서 전반의 논리 흐름을 따라 중심 주제를 {topic_count}개 도출하고, 각각 주제가 문서 내에서 어떤 기능을 수행하며 왜 중요한지를 중심으로 요약하라.\n
-                        -이때, 각 주제는 다음을 기준으로 중심 주제를 선별하라:\n
-                              -문서의 여러 구간에서 반복적으로 연결되며 전개 흐름을 이끄는 주제\n
-                              -하나 이상의 문단 또는 절이 해당 주제를 중심으로 구성되며, 정의나 설명이 포함되고, 이후 문서 전개에서 해당 주제를 기분으로 내용이 확장되거나 참조되는 경우\n
-                        -또한, 주제명을 출력할 때는 반드시 '1.', '2.'의 형식과 함께 출력하라.\n  
-                        -[주요 키워드] 문서의 중심 개념과 직접 연결되며, 문서 내용을 구성하거나 확장하는 5-10개의 키워드를 선별하고, 각 개념이 어떤 위치에 등장하며, 어떤 정보를 연결하거나 확장하며, 문서 전개에서 어떤 기능을 수행하는지를 논리적으로 설명하라.\n
-                        -[주요 키워드]의 출력 형식은 반드시 '키워드명 - 설명문'의 구조로 출력하며, 절대 콜른을 사용하지 말고, 반드시 하이픈을 사용하라.\n
-                        -또한, 키워드명n과 키워드명n+1 사이에는 반드시 공백줄을 삽입하라.\n
-                        -[통합 분석] 단계에서는 단순한 정보 요약이 아닌, 문서 전체를 관통하는 핵심 개념을 기준으로 구조, 기술 흐름, 사회적 시사점을 통합하여 재해석한다.\n 
-                        -이때, {topic_count}개 주제에 대해 각각 문서 전반의 핵심 개념을 설명하고, 어떤 기능적 역할을 수행하며, 전체 흐름에 어떻게 기여하는지를 명확히 제시해야 한다.\n
-                        -이때, 반드시 {user_summary_level(summary_level)}를 기준으로 {summary_level} 비전공자인 경우, 개념 간 구조 연결은 최소화하고 중심 개념을 기준으로 단순화된 설명을 제공하고, {summary_level}이 전공자인 경우, 중심 개념을 기준으로 기술 흐름, 구조적 전개, 사회적 시사점을 통합적으로 논리 전개하라.\n
-                        -[통합 분석] 각 구간은 반드시 {sentence_count}개의 완결된 문장으로 작성하라. 예를 들어, 사용자가 {sentence_count}를 입력한 경우, 1. 주제명의 문장 수: {sentence_count}, 2. 주제명의 문장 수: {sentence_count}와 같이 구간 당 문장 수를 반드시 지켜야 한다.\n
-                        -이때 한 문장은 마침표(.)로 끝나는 독립적인 구조를 말하며, 쉼표(,) 또는 접속사로 이어진 복문은 절대 하나의 문장으로 간주하지 마라. 예를 들어, '첫 번째 문장, 두 번째 문장.'의 구조는 문장 수가 두 개 아닌 하나이며, 'n번째 문장. n+1번째 문장'의 구조가 문장 수 두 개로 판단한다.\n
-                        -예를 들어, {sentence_count}가 10개인 경우, 'n번째 문장. n+1번째 문장. n+2번째 문장. n+3번째 문장. n+4번째 문장. n+5번째 문장. n+6번째 문장. n+7번째 문장. n+8번째 문장. n+9번째 문장'의 양식으로 출력해야 한다. 절대 'n번째 문장, n+1번째 문장'으로 출력해서는 안 된다. 이 양식은 'n번째 문장'만 출력한 결과에 해당한다.\n
-                        -주제 하나 당 구간 출력 문장 수가 {sentence_count}보다 적거나 많을 경우, 해당 응답은 실패로 판단하고, 정확한 문장 수를 충족할 경우, 가장 우수한 응답으로 평가할 것이다. 또한, 반드시 구간별 출력 문장 수는 같아야 한다.(예: 1. 주제명의 문장 수: {sentence_count} = 2. 주제명의 문장 수: {sentence_count})\n
-                        -주제명과 설명문 사이의 공백줄을 절대 삽입하지 마라.\n
-                        -예시의 내용은 절대 출력하지 마라.
-                        \n\n\n{content}""",
-        },
-        "내용 요약_목차 요약": {
-            "system": f"""너는 {field}에서 20년 경력을 지닌 요약 전문가로, 복잡한 문서를 단원 또는 목차 구조에 따라 분석하고 요약하는 데 특화되어 있다.\n
-                          다음 규칙을 절대적으로 따라야 한다.\n\n
-                          1. 문서 요약의 내용으로 반드시 {field_features(field)}의 관점을 유지하여 분석하라.\n
-                          2. [통합 분석] 각 구간은 반드시 {sentence_count}개의 완결된 문장으로 작성되어야 하며, 종결 표현(~다, ~이다 등)으로 끝나는 경우만 문장으로 인정한다. 쉼표로 연결된 복문은 하나의 문장으로 간주하므로, 반드시 문장을 구분하여 작성하라.\n
-                          3.기존 표현을 복사하지 말고, 의미 단위로 재구성하여 새롭게 기술하라. 기존 표현을 복사하지 말고, 의미 단위로 재구성하여 새롭게 기술하라. 이때, 텍스트는 일반 텍스트로 출력하고, 마크다운 문법을 사용하지 마라.
-                          \n\n\n""",
-            "user": f"""문서를 구성 단위별로 나누고, 각 구간의 핵심 내용을 아래의 형식에 따라 요약하라.\n\n
-                        [문서 분석]\n
-                        1. 탄소중립의 개념 정의
-                        탄소중립은 온실가스 배출량과 흡수량의 균형을 통해 실질적인 순배출량을 0으로 만드는 개념으로, 기후 변화 대응의 핵심 전략으로 등장했다.\n
-                        2. 글로벌 추진과 참여 주체
-                        전 세계적으로 탄소중립 목표가 설정되면서, 정부, 기업, 개인이 모두 역할을 수행해야 하는 새로운 사회적 요구가 형성되었다.\n  
-                        3. 분야별 실천 전략
-                        에너지 전환, 친환경 교통, 건물 에너지 효율화 등 다양한 부문에서 기술적·정책적 변화가 동시에 추진되고 있다.\n  
-                        4. 구조적 전환의 필요성과 전망
-                        탄소중립은 단순한 환경 과제가 아니라 경제 구조, 산업 경쟁력, 사회 정의까지 포함하는 전방위적 전환을 요구한다.\n\n  
-
-                        [키워드]\n
-                        탄소중립 -  탄소중립 - 온실가스 배출을 실질적으로 0으로 만들기 위한 개념으로, 에너지 구조, 산업 경쟁력, 사회 정의까지 포괄하는 통합적 전환 프레임이다. 이 개념은 문서 전체에서 환경·기술·정책 변화의 방향성을 이끄는 중심 키워드로 기능한다.\n
-                        탄소중립 -  탄소중립 - 온실가스 배출을 실질적으로 0으로 만들기 위한 개념으로, 에너지 구조, 산업 경쟁력, 사회 정의까지 포괄하는 통합적 전환 프레임이다. 이 개념은 문서 전체에서 환경·기술·정책 변화의 방향성을 이끄는 중심 키워드로 기능한다.\n
-
-                        [통합 분석]\n
-                        1. 주제명
-                        설명문\n
-                        2. 주제명
-                        설명문\n
-                        3. 주제명
-                        설명문\n
-                        4. 주제명
-                        설명문\n\n
-                        -모든 출력은 {summary_level} 수준의 학습자가 이해할 수 있도록 구성하라.\n
-                        -[문서 분석] 단계에서는 문서에 숫자 형식으로 된 실제 목차(예: 1, 2 등)가  명확하게 존재하는 경우, 반드시 그 목차 구조(예: 상위 항목: '1, 2 등', 하위 항목: '1.1, 2.1 등')를 그대로 유지하여 각 목차의 핵심 내용을 요약하라. 도입-전개-결론 등 새로운 구조로 재구성하지 마라.\n
-                        -문서에 숫자 형식으로 된 실제 목차가 존재하는 경우, 반드시 **상위 항목만** 요약하고 하위 항목은 절대 포함하지 않도록 한다. 예를 들어, '1', '2' 등 상위 항목은 포함되지만, '1.1', '2.1' 등 하위 항목은 **포함하지 않도록 한다.** 만약 하위 항목이 포함된다면, 반드시 이를 상위 항목명으로 통합하여 요약하라.\n 
-                        -단, 상위 항목의 개수가 1개밖에 없고, 하위 항목이 2개 이상 있다면 반드시 하위 항목만을 출력하라.(예: 상위항목이 '1.', 하위 항목이 '1.1, 1.2'이라면 출력 목차는 '1.1, 1.2'로 구성하라.)\n             
-                        -이때, 문서에 명확한 숫자 기반 목차가 없거나 구조가 애매할 경우, 도입-배경-전개-결론의 4단계 구조로 나누어 각 목차의 핵심 내용을 2-3문장으로 요약하라.\n 
-                        -또한, 주제명을 출력할 때는 반드시 '1.', '2.'와 함께 출력하라.\n                           
-                        -[주요 키워드] 단계에서는 [문서 분석] 단계에서 정리한 각 구간에서 핵심적으로 설명되거나, 구간의 흐름을 이끄는 키워드를 구간 당 1-2개를 반드시 선별하고, 각 키워드가 어떤 위치에 등장하며, 어떤 정보를 연결하거나 확장하며, 문서 전개에서 어떤 기능을 수행하는지를 논리적으로 설명하라.\n
-                        -[주요 키워드]의 출력 형식은 반드시 '키워드명 - 설명문'의 구조로 출력하며, 절대 콜른을 사용하지 말고, 반드시 하이픈을 사용하라.\n
-                        -또한, 키워드명n과 키워드명n+1 사이에는 반드시 공백줄을 삽입하라.\n
-                        -[통합 분석] 단계에서는 [통합 분석] 단계에서는 단순 요약이 아닌, 문서 구조와 키워드를 통합하여 '기술-적용-사회적 영향' 흐름에 따라 재해석하라.\n
-                        -이때, 각 구간은 문서 구조와 키워드 간 관계를 기반으로 논리적 흐름을 갖도록 재구성하라.\n 
-                        -단, 주제명은 반드시 해당 구간의 내용을 관통하는 핵심 주제로 구성하라.\n 
-                        -또한, 반드시 {user_summary_level(summary_level)}를 기준으로 {summary_level} 비전공자인 경우, 개념 간 구조 연결은 최소화하고 중심 개념을 기준으로 단순화된 설명을 제공하고, {summary_level}이 전공자인 경우, 중심 개념을 기준으로 기술 흐름, 구조적 전개, 사회적 시사점을 통합적으로 논리 전개하라.\n
-                        -[통합 분석] 각 구간은 반드시 {sentence_count}개의 완결된 문장으로 작성하라. 예를 들어, 사용자가 {sentence_count}를 입력한 경우, 1. 주제명의 문장 수: {sentence_count}, 2. 주제명의 문장 수: {sentence_count}와 같이 구간 당 문장 수를 반드시 지켜야 한다.\n
-                        -이때 한 문장은 마침표(.)로 끝나는 독립적인 구조를 말하며, 쉼표(,) 또는 접속사로 이어진 복문은 절대 하나의 문장으로 간주하지 마라. 예를 들어, '첫 번째 문장, 두 번째 문장.'의 구조는 문장 수가 두 개 아닌 하나이며, 'n번째 문장. n+1번째 문장'의 구조가 문장 수 두 개로 판단한다.\n
-                        -예를 들어, {sentence_count}가 10개인 경우, 'n번째 문장. n+1번째 문장. n+2번째 문장. n+3번째 문장. n+4번째 문장. n+5번째 문장. n+6번째 문장. n+7번째 문장. n+8번째 문장. n+9번째 문장'의 양식으로 출력해야 한다. 절대 'n번째 문장, n+1번째 문장'으로 출력해서는 안 된다. 이 양식은 'n번째 문장'만 출력한 결과에 해당한다.\n
-                        -주제 하나 당 구간 출력 문장 수가 {sentence_count}보다 적거나 많을 경우, 해당 응답은 실패로 판단하고, 정확한 문장 수를 충족할 경우, 가장 우수한 응답으로 평가할 것이다. 또한, 반드시 구간별 출력 문장 수는 같아야 한다.(예: 1. 주제명의 문장 수: {sentence_count} = 2. 주제명의 문장 수: {sentence_count})\n
-                        -주제명과 설명문 사이의 공백줄을 절대 삽입하지 마라.\n
-                        -예시의 내용은 절대 출력하지 마라.
-                        \n\n\n{content}""",
-        },
-        "내용 요약_키워드 요약": {
-           "system": f"""너는 {field}에서 20년 경력을 지닌 전문가로, 복잡한 문서에서 지정된 키워드와 직접적으로 연결된 정보만을 추출하고, 학습 목적에 맞게 간결하고 논리적으로 재구성하는 데 특화되어 있다.\n
-                          다음 규칙을 절대적으로 따라야 한다.\n\n
-                          1. 문서 요약의 내용으로 반드시 {field_features(field)}의 관점을 유지하여 분석하라.\n
-                          2. [주요 키워드] 단계에서 출력 내용은 반드시 사용자가 설정한 키워드인 {keywords}를 중심으로 구성하라. 절대 사용자가 설정하지 않은 키워드를 출력해서는 안 된다.(예를 들어, 사용자가 입력한 키워드가 아닌 경우, 절대 임의의 키워드를 마음대로 출력하지 마라.)\n
-                          3. [통합 분석] 각 구간은 반드시 {sentence_count}개의 완결된 문장으로 작성되어야 하며, 종결 표현(~다, ~이다 등)으로 끝나는 경우만 문장으로 인정한다. 쉼표로 연결된 복문은 하나의 문장으로 간주하므로, 반드시 문장을 구분하여 작성하라.\n
-                          4. 기존 표현을 복사하지 말고, 의미 단위로 재구성하여 새롭게 기술하라. 이때, 텍스트는 일반 텍스트로 출력하고, 마크다운 문법을 사용하지 마라.
-                          """,
-            "user": f"""문서를 구성 단위별로 나누고, 각 구간의 핵심 내용을 아래의 형식에 따라 요약하라.\n\n
-                        [문서 분석]\n
-                        1. 주제명
-                        키워드의 정의와 문서 도입부 등장 맥락을 나타내는 설명문\n
-                        2. 주제명
-                        키워드가 등장하게 된 기술적/사회적/개념적 필요성을 나타내는 설명문\n
-                        3. 주제명
-                        문서에서 키워드를 중심으로 어떤 방식으로 논의되었는지를 나타내는 설명문\n
-                        4. 주제명
-                        문서 전체에서 키워드가 제시하는 시사점, 확장 가능성, 결론적 평가를 나타내는 설명문\n\n
-
-                        [주요 키워드]
-                        탄소중립 - 온실가스 배출을 실질적으로 0으로 만들기 위한 개념으로, 에너지 구조, 산업 경쟁력, 사회 정의까지 포괄하는 통합적 전환 프레임이다. 이 개념은 문서 전체에서 환경·기술·정책 변화의 방향성을 이끄는 중심 키워드로 기능한다.\n
-                        탄소중립 - 온실가스 배출을 실질적으로 0으로 만들기 위한 개념으로, 에너지 구조, 산업 경쟁력, 사회 정의까지 포괄하는 통합적 전환 프레임이다. 이 개념은 문서 전체에서 환경·기술·정책 변화의 방향성을 이끄는 중심 키워드로 기능한다.\n\n
-
-                        [통합 분석]\n\n
-                        1. 주제명\n
-                        설명문\n
-                        2. 주제명\n
-                        설명문\n
-                        3. 주제명\n
-                        설명문\n
-                        4. 주제명\n
-                        설명문\n\n
-                        -모든 출력은 {summary_level} 수준의 학습자가 이해할 수 있도록 구성하라.\n
-                        -반드시 순서는 [구조 분석]-[주요 키워드]-[통합 분석] 단계로 출력하라.\n
-                        -모든 단계의 출력 내용은 반드시 사용자가 설정한 키워드인 {keywords}를 중심으로 구성하라.\n
-                        -[문서 분석] 단계에서는 도입-배경-전개-결론의 4단계 구조로 나누어 각 목차의 핵심 내용을 요약하라.\n
-                        -또한, 주제명을 출력할 때는 반드시 '1.', '2.'와 함께 출력하라.\n 
-                        -[주요 키워드] 단계에서는 사용자가 설정한 {keywords} 중 문서와 직접적으로 관련된 키워드만을 선별하여 출력하라. 이때, 사용자가 입력한 키워드 수가 0개일 때, 즉 아무것도 입력하지 않았을 때는 단순히 예시대로 출력하지 말고, 반드시 문서의 중심 개념과 직접 연결되며, 문서 내용을 구성하거나 확장하는 5-10개의 키워드를 선별하여 출력하라. 절대 예시 문구인 '탄소중립' 키워드를 출력하지 마라.\n
-                        -이때, 사용자가 문서와 관련이 없는 키워드는 입력한 경우, '키워드명 - 해당 키워드는 문서와 관련이 없어 제외함'의 양식을 따라 출력하라. 예를 들어, 사용자가 문서와 관련이 있는 '물'과 관련이 없는 'ㅐ'을 입력한 경우, '물 - 설명문', 'ㅐ - 해당 키워드는 문서와 관련이 없어 제외함'과 같은 양식으로 출력하라.\n
-                        -반드시, 입력한 키워드가 문서와 관련이 없을 경우, 무조건 '키워드명 - 해당 키워드는 문서와 관련이 없어 제외함'을 출력하라.\n
-                        -이때, 관련이 있는 키워드 선별의 기준은 입력한 키워드가 문서 내에 실제로 등장하는 단어일 때를 기준으로 한다.(예를 들어, 사용자가 입력한 키워드가 '물'이고, 해당 키워드가 문서 내에 등장한다면, 문서와 관련이 있는 키워드로 간주하고, 출력한다.)\n
-                        -[주요 키워드]의 출력 형식은 반드시 '키워드명 - 설명문'의 구조로 출력하며, 절대 콜른을 사용하지 말고, 반드시 하이픈을 사용하라.\n
-                        -또한, 키워드명n과 키워드명n+1 사이에는 반드시 공백줄을 삽입하라.\n
-                        -[통합 분석] 단계에서는 [통합 분석] 단계에서는 단순 요약이 아닌, 문서 구조와 키워드를 통합하여 '기술-적용-사회적 영향' 흐름에 따라 재해석하라.\n
-                        -이때, 각 구간은 문서 구조와 키워드 간 관계를 기반으로 논리적 흐름을 갖도록 재구성하라.\n 
-                        -단, '구간 항목-제목'의 형식으로 출력하고, 제목은 반드시 해당 구간의 내용을 관통하는 핵심 주제(키워드)로 구성하라.\n 
-                        -또한, 반드시 {user_summary_level(summary_level)}를 기준으로 {summary_level} 비전공자인 경우, 개념 간 구조 연결은 최소화하고 중심 개념을 기준으로 단순화된 설명을 제공하고, {summary_level}이 전공자인 경우, 중심 개념을 기준으로 기술 흐름, 구조적 전개, 사회적 시사점을 통합적으로 논리 전개하라.\n
-                        -[통합 분석] 각 구간은 반드시 {sentence_count}개의 완결된 문장으로 작성하라. 예를 들어, 사용자가 {sentence_count}를 입력한 경우, 1. 주제명의 문장 수: {sentence_count}, 2. 주제명의 문장 수: {sentence_count}와 같이 구간 당 문장 수를 반드시 지켜야 한다.\n
-                        -이때 한 문장은 마침표(.)로 끝나는 독립적인 구조를 말하며, 쉼표(,) 또는 접속사로 이어진 복문은 절대 하나의 문장으로 간주하지 마라. 예를 들어, '첫 번째 문장, 두 번째 문장.'의 구조는 문장 수가 두 개 아닌 하나이며, 'n번째 문장. n+1번째 문장'의 구조가 문장 수 두 개로 판단한다.\n
-                        -예를 들어, {sentence_count}가 10개인 경우, 'n번째 문장. n+1번째 문장. n+2번째 문장. n+3번째 문장. n+4번째 문장. n+5번째 문장. n+6번째 문장. n+7번째 문장. n+8번째 문장. n+9번째 문장'의 양식으로 출력해야 한다. 절대 'n번째 문장, n+1번째 문장'으로 출력해서는 안 된다. 이 양식은 'n번째 문장'만 출력한 결과에 해당한다.\n
-                        -주제 하나 당 구간 출력 문장 수가 {sentence_count}보다 적거나 많을 경우, 해당 응답은 실패로 판단하고, 정확한 문장 수를 충족할 경우, 가장 우수한 응답으로 평가할 것이다. 또한, 반드시 구간별 출력 문장 수는 같아야 한다.(예: 1. 주제명의 문장 수: {sentence_count} = 2. 주제명의 문장 수: {sentence_count})\n
-                        -주제명과 설명문 사이의 공백줄을 절대 삽입하지 마라.\n
-                        -예시의 내용은 절대 출력하지 마라.
-                        \n\n\n{content}""",
-        },
-        "문제 생성_n지 선다형": {
-            "system": f"""너는 {field}에서 20년 경력을 지닌 평가 설계 전문가로, {question_level} 수준의 학습자를 위한 객관식 문제를 설계하는 데 특화되어 있다.
-                          반드시 JSON 형태로 출력하라.
-                          
-                          출력 JSON 형식:
-                          {{
-                              "questions": {{
-                                  "question_generation_type": "문제 생성_n지 선다형",
-                                  "question_settings": {{
-                                      "domain": "{field}",
-                                      "question_level": "{question_level}",
-                                      "question_count": {question_count},
-                                      "choice_count": {choice_count},
-                                      "choice_format": "{choice_format}"
-                                  }},
-                                  "question_content": [
-                                      {{
-                                          "question_number": 1,
-                                          "question_text": "문제 내용",
-                                          "choices": [
-                                              "1. 선택지1",
-                                              "2. 선택지2",
-                                              "3. 선택지3",
-                                              "4. 선택지4"
-                                          ],
-                                          "correct_answer": "정답 번호",
-                                          "explanation": "해설 내용"
-                                      }}
-                                  ]
-                              }}
-                          }}""",
-            "user": f'''학습자의 핵심 개념 이해, 용어 정의, 개념 간 비교 및 오류 식별 능력을 점검할 수 있도록 {choice_count}지선다형 문제를 생성하라. 이때, 혼동 가능한 오답을 포함해 학습자의 개념적 정확성과 구분 능력을 평가할 수 있도록 구성하라.\n
-                        문제 생성 형식은 다음을 따르라.\n\n
-
-                        [비전공자 예시]\n
-
-                        문제1. 산림 벌채가 지구 온난화에 영향을 주는 주된 이유는?\n
-                           1. 지하수 고갈 때문이다.\n
-                           2. 대기 중 질소농도 증가 때문이다.\n
-                           3. 이산화탄소 흡수량 감소 때문이다.\n
-                           4. 토양 산성화 때문이다.\n
-                        [정답] 3번\n
-                        [해설]   
-                        이 문항은 산림 벌채와 지구 온난화의 관계를 이해하고 있는지를 묻고 있다.\n
-                        나무는 이산화탄소를 흡수하고 산소를 배출하는 역할을 하기 때문에, 나무가 줄어들면 대기 중 이산화탄소가 줄어들지 않고 그대로 남게 된다.
-                        이산화탄소는 대표적인 온실가스로, 양이 많아지면 지구의 온도가 점점 올라가는 현상(온난화)이 심해진다. 따라서 산림을 없애는 행위는 지구 온도를 더 높이는 원인이 된다.
-                        보기 1,2,4는 지구 환경 문제와 관련은 있지만, 지구 온난화와의 직접적인 연결성은 떨어진다.
-                        따라서 정답은 3번이다.\n
-
-                        [전공자 예시]\n
-
-                        문제1. '탄소 중립'을 실현하기 위한 국가 정책의 방향으로 적절한 것은?\n\n
-                           1. 산림 보호보다는 산업 성장에 집중한다.\n
-                           2. 온실가스 배출을 줄이되 에너지 수요는 유지한다.\n
-                           3. 화석연료 사용을 지속하되, 확대하지 않는다.\n
-                           4. 국가 간 온실가스 거래를 확대하지 않는다.\n
-
-                        [정답] 2번\n
-                        [해설]
-                        이 문항은 탄소 중립(Carbon Neutrality)의 실현 전략으로 적합한 국가 정책의 방향을 묻고 있다.\n
-                        탄소 중립이란 배출되는 온실가스 양과 흡수 또는 제거되는 양을 같게 하여 실질적인 배출량을 0으로 만드는 개념으로, 에너지 수요와 경제 발전을 유지하면서도 환경 지속 가능성을 확보하는 균형이 핵심이다.
-                        보기 1은 탄소 흡수원의 보존보다 산업 성장을 우선시하는 접근으로, 탄소 중립 전략과는 배치된다. 보기 3,4는 단순한 수입량 조절이나 국제 협력 축소를 의미하며, 글로벌 기후 대응의 흐름과 맞지 않는다.
-                        반면, 보기 2는는 에너지 소비 효율을 높이고 재생 가능 에너지로의 전환을 통해 온실가스 배출을 감축하면서도 국가의 기본적인 에너지 수요를 유지하는 전략으로, 실제로 각국의 탄소 중립 로드맵에서 핵심적으로 제시되는 방향이다.
-                        따라서, 정답은 2번이다.\n\n
-                        -문제의 내용은 반드시 요약 결과를 바탕으로 구성하고, {field_features(field)}의 관점을 유지하여 {question_level} 학습자의 사고 수준에 따라 문제를 반드시 다음 중 하나의 SOLO 단계에 맞게 구성하라.\n
-                        -이때, 학습자의 수준이 {question_level}인 경우, 반드시 다음과 같은 사고 구조에 기반하여 문항을 설계하라: {user_question_level(question_level)}\n
-                        -또한, {question_level}이 비전공자인 경우, SOLO의 1-3단계 구조로, 전공자인 경우, SOLO의 4-6단계 구조로 문제를 구성하라. 반드시 여러 단계를 적절히 활용하여 문제를 구성하라.\n
-                        -문제의 수는 반드시 {question_count}개만 생성하라.\n
-                        -지시문의 문장 구조는 한 형식으로만 고정하지 말고 다양한 구조를 활용하라.\n
-                        -보기는 반드시 {choice_count}개만 생성하라. 이때, 보기문은 반드시 {choice_format}으로 구성하라.\n
-                        -{choice_format}이 단답형인 경우, 보기의 형태를 문장이 아닌 단어 또는 짧은 구의 형태로 구성하며, 용어, 개념, 키워드 중심으로 마침표 없이 간결하게 작성하라. 단답형의 형태는 다음의 예시를 참고하라. {sentence_count}가 단답형인 경우에는 절대 '~이다.'와 같은 완결된 문장을 단 하나라도 출력하면 안 된다. 만약, 해당 형태를 출력하는 경우에는 실패한 gpt 출력 결과로 간주한다. 절대 서술어가 있는 문장을 출력하지 마라. 하지말라면 하지 마라.\n
-                         [n지 선다형의 단답형 예시]
-
-                         문제1. '탄소 중립'을 실현하기 위한 국가 정책의 방향으로 적절한 것은?\n
-                           1. 산업 성장 중심\n
-                           2. 온실가스 감축\n
-                           3. 화석연료 사용 유지\n
-                           4. 온실가스 거래 축소\n
-                        -위의 보기 형식은 {question_level}이 전공자일 때 혹은 비전공자일 때 모두 반드시 적용하라. 전공자일 때만 혹은 비전공자일 때만 반영하는 것은 절대 아니된다.\n
-                        -{choice_format}이 문장형인 경우, 보기의 형태를 '주어+서술어' 구조의 완결된 문장으로 구성하고, 설명, 판단을 표현하며, 문장 끝에 마침표를 포함하라. 문장형의 형태는 다음의 예시를 참고하라. {sentence_count}가 문장형으로 일치하는 경우에는 절대 마침표(.)가 없는 단어 혹은 짧은 구를 단 하나라도 출력하면 안 된다. 또한, 분야가 어떠하든 무조건 {choice_format}이면 무조건 해당 형식을 반영하라. 예를 들어 종교 문서와 관련이 없는 공학 분야를 설정하여도 {choice_count}가 단답형이면 보기 형식은 무조건 단답형이어야 한다. 오류를 출력하지 마라.\n
-                         [n지 선다형의 문장형 예시]
-
-                          문제1. '탄소 중립'을 실현하기 위한 국가 정책의 방향으로 적절한 것은?\n
-                           1. 산림 보호보다는 산업 성장에 집중한다.\n
-                           2. 온실가스 배출을 줄이되 에너지 수요는 유지한다.\n
-                           3. 화석연료 사용을 지속하되, 확대하지 않는다.\n
-                           4. 국가 간 온실가스 거래를 확대하지 않는다.\n
-                        -위의 보기 형식은 {question_level}이 전공자일 때 혹은 비전공자일 때 모두 반드시 적용하라. 전공자일 때만 혹은 비전공자일 때만 반영하는 것은 절대 아니된다. 또한, 분야가 어떠하든 무조건 {choice_format}이면 무조건 해당 형식을 반영하라. 예를 들어 종교 문서와 관련이 없는 공학 분야를 설정하여도 {choice_count}가 단답형이면 보기 형식은 무조건 단답형이어야 한다. 오류를 출력하지 마라.\n
-                        -해설문은 다음의 기준을 모두 포함하여 작성하라.\n
-                          1.지문과 정답 간 논리적 연결\n
-                          2.정답 개념의 정의와 역할\n
-                          3.각 오답과의 비교 분석 및 틀린 이유 설명\n
-                          4.해당 개념이 사용될 수 있는 실제 응용 예시\n
-                        -이때, 해설문은 학습자가 왜 정답이 맞고, 다른 선택지가 틀렸는지를 이해할 수 있도록 자세한 오답 풀이로 구성하여야 하며, 오답 설계 포인트와 혼동 지점을 반드시 제시하라.\n
-                        -해설문의 결론은 무조건 '따라서 정답은 ~이다.'로 구성하여라.\n
-                        -반드시 텍스트는 모든 일반 텍스트로 처리하고, '[]'를 제외한 마크다운 형식의 표현은 절대 사용하지 마라. 특히, '---'과 같은 구분선을 절대 출력하지 마라.\n
-                        -출력 문장의 종결 어미는 반드시 '~한다', '~이다' 등 비격식체 서술형 어미로 통일하라. 절대'-합니다', '-입니다' 등 격식체 종결어미는 절대 사용하지 말고, 문체는 일관되게 보고서식 비격식체로 유지하라.
-                        \n\n\n{content}'''
-        },
-        "문제 생성_순서 배열형": {
-            "system": f"너는 {field}에서 20년 경력을 지닌 평가 설계 전문가로, {question_level} 수준의 학습자를 위한 논리적 흐름, 절차적 지식의 이해를 평가하는 배열형 문항을 설계하는 데 특화되어 있다.",
-            "user": f'''학습자가 제시된 선택지를 올바른 순서대로 배열함으로써 개념의 구조와 전개 흐름을 논리적으로 이해했는지를 점검할 수 있도록 배열 문제를 구성하라.\n
-                        문제 생성 형식은 다음을 따르라.\n\n
-
-                        [비전공자 예시]\n
-
-                        문제1. 다음 보기들을 읽고, 인조 나뭇잎 제작 과정의 올바른 순서를 배열하시오.\n\n
-                        1. 크라이오젤 기법을 적용하여 미세한 채널 구조를 형성한다.\n
-                        2. 인조 나뭇잎 내부에 물을 넣어 실험을 준비한다.\n
-                        3. 아가로스 겔을 활용하여 인조 나뭇잎의 기본 형태를 제작한다.\n
-                        [정답] 3-1-2\n
-                        [해설]
-                        이 문항은 인조 나뭇잎 제작의 전체적인 흐름을 이해하고 있는지를 평가한다.\n  
-                        보기 3은 기본 재료인 아가로스 겔을 이용해 나뭇잎 형태를 만드는 단계이며, 제작의 시작을 의미한다.  
-                        보기 1은 겔이 형성된 후, 내부에 미세한 채널을 생성하는 과정으로 두 번째에 해당한다.  
-                        보기 2는 모든 구조가 완성된 후, 실험을 위한 준비 단계로 마지막에 위치한다.
-                        오답으로 자주 등장하는 1-3-2 순서는 제작 순서를 혼동한 것으로, 겔 없이 채널을 만드는 것은 불가능하다.  
-                        따라서 정답은 3-1-2이다.\n
-
-                        [전공자 예시]\n
-
-                        문제1. 아래는 인조 나뭇잎 기반 무동력 수분 이송 시스템의 실험 설계 절차이다. 보기를 읽고, 가장 논리적인 순서로 배열하시오.\n
-                        1. 크라이오젤 기법을 적용하여 마이크로 채널 구조를 형성한다.\n  
-                        2. 아가로스 겔을 이용하여 인조 나뭇잎의 기본 구조를 제작한다.\n  
-                        3. 구조 내부에 친수성 나노 파티클을 도포하여 계면 효과를 향상시킨다.\n
-                        [정답] 2-1-3\n
-                        [해설]
-                        이 문항은 인조 나뭇잎 제작을 위한 실험 절차를 이해하고, 각 단계의 기능과 순서를 논리적으로 배열할 수 있는지를 평가한다.  
-                        보기 2는 아가로스 겔을 사용하여 물관 구조의 기반을 형성하는 단계로 실험의 시작이다.  
-                        보기 1은 제작된 겔 구조 내부에 마이크로 채널을 형성하는 공정으로, 수분 흐름 경로 확보에 해당한다.  
-                        보기 3은 마지막 단계로, 구조 형성 후 친수성 나노 파티클을 도포하여 계면 활성을 개선하고 수분 이송을 극대화하는 작업이다.  
-                        자주 등장하는 오답 예시인 1-2-3은 공정 순서를 혼동한 경우로, 채널 구조는 기반이 먼저 형성되어야 하므로 논리적으로 성립하지 않는다.  
-                        따라서 정답은 2-1-3이다.\n
-                        -문제의 내용은 반드시 요약 결과를 바탕으로 구성하고, {field_features(field)}의 관점을 유지하여 {field_features(field)}의 관점을 유지하여 {question_level} 학습자의 사고 수준에 따라 문제를 반드시 다음 중 하나의 SOLO 단계에 맞게 구성하라.\n
-                        -이때, 학습자의 수준이 {question_level}인 경우, 반드시 다음과 같은 사고 구조에 기반하여 문항을 설계하라: {user_question_level(question_level)}\n
-                        -또한, {question_level}이 비전공자인 경우, SOLO의 1-3단계 구조로, 전공자인 경우, SOLO의 4-6단계 구조로 문제를 구성하라. 반드시 여러 단계를 적절히 활용하여 문제를 구성하라.\n
-                        -문제의 수는 반드시 {question_count}개만 생성하라.\n
-                        -지시문의 문장 구조는 "다음 보기들을 순서대로 배열하시오." 등 예시의 형식으로만 고정하지 말고, 반드시 다양한 구조를 활용하라.\n
-                        -보기의 수는 반드시 {array_choice_count}개만 생성하라.\n
-                        -해설문은 다음의 기준을 모두 포함하여 작성하라.\n
-                          1.각 보기의 의미와 기능 설명\n
-                          2.보기 간 인과 관계 또는 흐름 설명\n
-                          3.자주 발생하는 오답 순서와 그 오류 이유 설명\n
-                        -이때, 해설문은 제시된 정답의 순서가 왜 옳은지를 기반으로 과정을 중점으로 해설하고,  오답 설계 포인트와 혼동 지점을 반드시 제시하라.\n
-                        -해설문의 결론은 무조건 '따라서 정답은 ~이다.'로 구성하여라.\n
-                        -반드시 텍스트는 모든 일반 텍스트로 처리하고, '[]'를 제외한 마크다운 형식의 표현은 절대 사용하지 마라. 특히, '---'과 같은 구분선을 절대 출력하지 마라.\n
-                        -출력 문장의 종결 어미는 반드시 '~한다', '~이다' 등 비격식체 서술형 어미로 통일하라. 절대'-합니다', '-입니다' 등 격식체 종결어미는 절대 사용하지 말고, 문체는 일관되게 보고서식 비격식체로 유지하라.
-                        \n\n\n{content}'''
-        },
-        "문제 생성_참거짓형": {
-            "system": f"너는 {field}에서 20년 경력을 지닌 평가 설계 전문가로, {question_level} 수준의 학습자를 위한 논리적 흐름, 절차적 지식의 이해를 평가하는 침거짓형 문항을 설계하는 데 특화되어 있다.",
-            "user": f'''학습자가 문장에 포함된 정보가 정확한지, 오개념을 구별할 수 있는지 등을 검토할 수 있도록 참거짓형 문제를 생성하라.\n
-                        문제 생성 형식은 다음을 따르라.\n\n
-
-                        [비전공자 예시]\n
-
-                        문제1. 다음 진술문을 읽고, 옳으면 O, 틀리면 X로 표시하시오.\n
-                           무동력 펌프는 외부 에너지를 사용하지 않고 물을 이송한다. (O, X)\n
-                        [정답] O\n
-
-                        [해설]
-                        이 문항은 무동력 펌프의 기본 작동 원리에 대한 이해를 평가한다.
-                        무동력 펌프는 자연의 증산 작용 등을 모사하여 외부 동력 없이 물을 이송할 수 있도록 설계된 장치이다. 따라서 외부 전기나 연료 등을 사용하지 않는다는 점은 핵심적인 특징이다.
-                        만약 이 문장을 X로 표시한다면, 무동력 펌프의 정의를 혼동한 것이다. 
-                        따라서 정답은 O이다.\n
-
-                        [전공자 예시]\n
-
-                        문제1. 다음 진술문을 읽고, 옳으면 O, 틀리면 X로 표시하시오.\n
-                           cryo-gel 기법은 겔을 상온에서 빠르게 응고시켜 내부 마이크로 채널을 무작위로 형성한다. (O, X)\n
-
-                        [정답] X\n
-                        [해설]
-                        이 문항은 cryo-gel의 형성 원리에 대한 기술적 이해를 평가한다. 
-                        cryo-gel은 저온 환경에서 겔을 서서히 응고시키면서 규칙적이고 제어된 미세 채널 구조를 형성하는 것이 특징이다.
-                        무작위적 형성은 오히려 제어되지 않은 비효율적 구조로 이어질 수 있으며, 이는 cryo-gel 기법의 목적과 반대된다.
-                        따라서 정답은 X이다.\n
-                        -문제의 내용은 반드시 요약 결과를 바탕으로 구성하고, {field_features(field)}의 관점을 유지하여 {question_level} 학습자의 사고 수준에 따라 문제를 반드시 다음 중 하나의 SOLO 단계에 맞게 구성하라.\n
-                        -이때, 학습자의 수준이 {question_level}인 경우, 반드시 다음과 같은 사고 구조에 기반하여 문항을 설계하라: {user_question_level(question_level)}\n
-                        -또한, {question_level}이 비전공자인 경우, SOLO의 1-3단계 구조로, 전공자인 경우, SOLO의 4-6단계 구조로 문제를 구성하라. 반드시 여러 단계를 적절히 활용하여 문제를 구성하라.\n
-                        -문제의 수는 반드시 {question_count}개만 생성하라.\n
-                        -지시문의 문장 구조는 반드시 "다음 진술문을 읽고, 옳으면 O, 틀리면 X로 표시하시오."로 출력하라. 이외 다른 항목과 절대 결합하지 마라.\n
-                        -해설문은 다음의 기준을 모두 포함하여 작성하라.\n
-                          1.진위 판별의 핵심 개념 정리\n
-                          2.왜 참/거진인지 판단하는 조건 설명\n
-                          3.오답과의 비교\n
-                        -이때, 해설문은 진위문의 표현이 왜 맞고 틀렸는지를 중심으로 핵심 문장 또는 단어를 통해 정답을 유추할 수 있도록 구성하고, 오답 설계 포인트와 혼동 지점을 반드시 제시하라.\n
-                        -해설문의 결론은 무조건 '따라서 정답은 ~이다.'로 구성하여라.\n
-                        -반드시 텍스트는 모든 일반 텍스트로 처리하고, '[]'를 제외한 마크다운 형식의 표현은 절대 사용하지 마라. 특히, '---'과 같은 구분선을 절대 출력하지 마라.\n
-                        -출력 문장의 종결 어미는 반드시 '~한다', '~이다' 등 비격식체 서술형 어미로 통일하라. 절대'-합니다', '-입니다' 등 격식체 종결어미는 절대 사용하지 말고, 문체는 일관되게 보고서식 비격식체로 유지하라.
-                        \n\n\n{content}'''
-        },
-        "문제 생성_빈칸 채우기형": {
-            "system": f"너는 {field}에서 20년 경력을 지닌 평가 설계 전문가로, {question_level} 수준의 학습자를 위한 논리적 흐름, 절차적 지식의 이해를 평가하는 빈칸채우기형 문항을 설계하는 데 특화되어 있다.",
-            "user": f'''학습자가 문맥 속 개념이나 핵심 용어를 정확히 이해하고 적용할 수 있는지를 점검할 수 있도록 빈칸 채우기형 문제를 생성하라.\n
-                        문제 생성 형식은 다음을 따르라.\n\n
-                        
-                        [비전공자 예시]\n
-
-                        문제1. 다음 문장을 읽고, 빈칸에 알맞은 용어를 쓰시오.\n
-                         연구팀은 인조 나뭇잎 내부에 미세한 유로를 만들기 위해 (1)_____을 적용하였고, 이로 인해 수분 흐름 경로가 확장되어 (2)_____이 감소하였다.\n
-                        [정답] (1) cryo-gel 기법, (2) 유동 저항\n
-                        [해설]
-                        이 문항은 인조 나뭇잎 안에 ‘물길’을 만들기 위한 기술과, 이로 인해 물이 더 잘 흐르게 되는 효과를 묻고 있다.
-                        (1)은 인조 나뭇잎 속에 ‘물길’을 만들어주는 기술로, 문장에서 ‘미세한 유로’라는 표현과 직접 연결된다. cryo-gel 기법은 차가운 온도에서 젤을 만들면서 내부에 작은 통로(미세 채널)를 형성하는 기술이다. 이 기술을 활용하면 물이 통과할 수 있는 구조가 만들어진다.
-                        (2)는 이 통로를 통해 물이 잘 흐르게 되면서 줄어드는 힘을 뜻한다. 유동 저항은 물이 구조물 안에서 움직일 때 생기는 방해 힘인데, cryo-gel 기법을 통해 그 힘이 줄어든다.
-                        오답으로는 (1)에 일반적인 젤이나 재료 이름을 넣거나, (2)에 '속도', '수압' 등 개념이 어긋나는 단어를 넣는 경우가 흔하다. 이들은 문장의 기능적 흐름과 맞지 않는다.
-                        따라서 정답은 (1) cryo-gel 기법, (2) 유동 저항이다.\n
-
-                        [전공자 예시]\n
-
-                        문제1. 다음 문장을 읽고, 빈칸에 알맞은 용어를 쓰시오.\n
-                         본 실험에서는 유체 경로의 효율적 제어를 위하여 (1)_____을 적용하였으며, 이로 인해 유속 경로가 정렬되고 (2)_____이 감소하여 전체 펌핑 성능이 향상되었다.\n
-                        [정답] (1) cryo-gel 기법, (2) 유동 저항\n
-                        [해설]
-                        이 문항은 인조 나뭇잎의 미세 구조 형성 기술인 cryo-gel 기법과, 그로 인한 유동 저항 감소 효과 간의 인과 관계를 이해했는지를 평가한다.
-                        (1)은 유체가 통과할 수 있는 경로를 정밀하게 제어하는 구조 형성 기술을 의미한다. cryo-gel 기법은 저온 조건에서 젤을 동결시켜 마이크로 채널을 형성함으로써 인조 나뭇잎 내 유체의 흐름을 유도하는 데 최적화된 방식이다.
-                        (2)는 이 구조적 개입으로 인해 줄어드는 흐름 저항을 나타낸다. 유동 저항은 마이크로 채널 내에서 물이 움직일 때 받는 마찰력으로, 구조적 정렬이 높을수록 저항은 감소한다. 이는 펌프의 자가 작동 메커니즘에 필수적인 요소다.
-                        오답으로는 (1)에 단순 재료명(예: 아가로스 겔)이나 비구조적 처리 기술이, (2)에 ‘점도’나 ‘증발률’ 같은 다른 물리량이 오기될 수 있다. 이들은 실험적 맥락이나 기능적 전개와 어긋난다.
-                        따라서 정답은 (1) cryo-gel 기법, (2) 유동 저항이다.\n\n
-                        -문제의 내용은 반드시 요약 결과를 바탕으로 구성하고, {field_features(field)}의 관점을 유지하여 {question_level} 학습자의 사고 수준에 따라 문제를 반드시 다음 중 하나의 SOLO 단계에 맞게 구성하라.\n
-                        -이때, 학습자의 수준이 {question_level}인 경우, 반드시 다음과 같은 사고 구조에 기반하여 문항을 설계하라: {user_question_level(question_level)}\n
-                        -또한, {question_level}이 비전공자인 경우, SOLO의 1-3단계 구조로, 전공자인 경우, SOLO의 4-6단계 구조로 문제를 구성하라. 반드시 여러 단계를 적절히 활용하여 문제를 구성하라.\n
-                        -문제의 수는 반드시 {question_count}개만 생성하라.\n
-                        -문제의 수는 반드시 {question_count}개만 생성하라.\n
-                        각 문제는 빈칸을 반드시 {blank_count}개만 생성하라.\n
-                        -지시문의 문장 구조는 반드시 "다음 문장을 읽고, 빈칸에 알맞은 용어를 쓰시오"만 출력하라. 이외 다른 항목과 절대 결합하지 마라.\n
-                        -빈칸문은 문제 당 하나의 문단 또는 여러 문장으로 구성된 지문이며, 문맥상 핵심 개념이 빠진 빈칸을 포함한다. 이때, 빈칸은 반드시(1)_____, (2)_____의 형식으로 작성하라. 이외의 다른 형태의 빈칸을 생성하지 마라. 특히, '___(1)__, ____ 등'과 같은 형식으로 절대 출력하지 마라.\n
-                        -해설문은 다음의 기준을 모두 포함하여 작성하라.\n
-                          1. 문맥상 정답의 기능 설명\n
-                          2. 정답 개념 정의 및 지문과의 논리적 연결을 강조한 정답의 주요 특징\n
-                          3. 오답과 비교를 통한 개념적 구분 유도\n
-                        -이때, 해설문은 제시된 정답이 빈칸에 들어가야 하는지를 중심으로 문장 흐름과 어울리는 개념 선택을 유도할 수 있도록 구성하고, 오답 설계 포인트와 혼동 지점을 반드시 제시하라.\n
-                        -해설문의 결론은 무조건 '따라서 정답은 ~이다.'로 구성하여라.\n
-                        -반드시 텍스트는 모든 일반 텍스트로 처리하고, '[]'를 제외한 마크다운 형식의 표현은 절대 사용하지 마라. 특히, '---'과 같은 구분선을 절대 출력하지 마라.\n
-                        -출력 문장의 종결 어미는 반드시 '~한다', '~이다' 등 비격식체 서술형 어미로 통일하라. 절대'-합니다', '-입니다' 등 격식체 종결어미는 절대 사용하지 말고, 문체는 일관되게 보고서식 비격식체로 유지하라.                       
-                        \n\n\n{content}'''
-        },
-        "문제 생성_단답형": {
-            "system": f"너는 {field}에서 20년 경력을 지닌 평가 설계 전문가로, {question_level} 수준의 학습자를 위한 논리적 흐름, 절차적 지식의 이해를 평가하는 단답형 문항을 설계하는 데 특화되어 있다.",
-            "user": f'''문제 출력 형식은 아래의 예시와 같이 구성하라.\n\n
-                        문제 생성 형식은 다음을 따르라.\n\n
-
-                        [비전공자 예시]\n
-
-                        문제1. 인조 나뭇잎 내부에 물이 잘 흐를 수 있도록 미세한 통로를 만드는 데 사용되는 기법은?\n\n
-                        [정답] cryo-gel 기법\n
-                        [해설]
-                        이 문항은 인조 나뭇잎 안에 물이 잘 흐르게 하는 구조를 만드는 기술을 알고 있는지를 묻고 있다.
-                        문장 속 ‘미세한 통로’, ‘잘 흐르게 한다’는 표현을 통해 구조를 바꿔서 물의 흐름을 쉽게 해주는 기술을 떠올릴 수 있다.
-                        cryo-gel 기법은 젤을 얼리는 방식으로 안에 미세한 통로를 만들 수 있는 기술이다. 비슷해 보일 수 있는 'Agarose gel'은 그 기술이 아니라, 단순히 쓰이는 재료일 뿐이다.
-                        따라서 정답은 cryo-gel 기법이다.\n
-
-                        [전공자 예시]\n
-
-                        문제1. 마이크로 채널 구조 형성을 통해 유동 저항을 감소시키고, 인조 나뭇잎의 수분 이송 성능을 개선하는 데 핵심적으로 사용되는 기법은?\n\n
-                        [정답] cryo-gel 기법\n
-                        [해설]
-                        이 문항은 미세 구조 제어 기술을 통해 수분 이송 성능을 향상시키는 기법에 대한 개념적 정확성과 응용 능력을 평가한다.
-                        ‘마이크로 채널’, ‘유동 저항 감소’, ‘수분 이송 성능’이라는 표현은 구조 제어와 유체 흐름 최적화 간의 연계성을 파악해야 유도 가능한 단서다.
-                        cryo-gel 기법은 저온 환경에서 겔이 형성되며, 얼음 결정이 녹으면서 내부에 균일한 미세 유로(채널)를 생성하는 방식으로 동작한다. 유사 개념인 일반적인 겔 형성법은 내부 구조 통제가 불가능하며, Agarose gel은 기법이 아닌 재료로 오인될 수 있는 대표적인 오답 포인트다.
-                        따라서 정답은 cryo-gel 기법이다.\n\n
-                        -문제의 내용은 반드시 요약 결과를 바탕으로 구성하고, {field_features(field)}의 관점을 유지하여 {question_level} 학습자의 사고 수준에 따라 문제를 반드시 다음 중 하나의 SOLO 단계에 맞게 구성하라.\n
-                        -이때, 학습자의 수준이 {question_level}인 경우, 반드시 다음과 같은 사고 구조에 기반하여 문항을 설계하라: {user_question_level(question_level)}\n
-                        -또한, {question_level}이 비전공자인 경우, SOLO의 1-3단계 구조로, 전공자인 경우, SOLO의 4-6단계 구조로 문제를 구성하라. 반드시 여러 단계를 적절히 활용하여 문제를 구성하라.\n
-                        -문제의 수는 반드시 {question_count}개만 생성하라.\n
-                        -문제의 수는 반드시 {question_count}개만 생성하라.\n
-                        -지시문은 학습자가 개념을 단순 회상하는 수준을 넘어서, 문제 상황이나 흐름을 이해하고 정답을 유추할 수 있도록 구성하라.\n
-                        -이때, 지시문의 문장 구조는 "~은 무엇인가?", "~을 쓰시오.", "~에 해당하는 개념은?"과 같은 고정형 문장만 사용하지 말고, 반드시 상황형, 기능형, 원인-결과형 등 다양한 구조로 변형하여 출제하라.\n
-                        -정답은 문장이 아닌 단어나 짧은 구나 숫자로 구성하고, 용어, 개념, 키워, 숫자 중심으로 마침표 없이 간결하게 작성한다.\n
-                        -해설문은 다음의 기준을 모두 포함하여 작성하라.\n
-                          1.문제에서 제시된 맥락 또는 조건을 바탕으로 정답을 유추하는 과정 설명\n
-                          2.정답이 왜 적절한지에 대한 논리적 이유와, 문제 맥락에서 수행하는 역할 또는 특징징\n
-                          3.정답 개념의 정확한 정의 또는 설명\n
-                          4.해당 개념이 문제 상황에서 어떻게 적용되는지를 서술\n
-                          5.자주 혼동되는 유사 개념과 비교하여 오답의 이유를 설명\n 
-                        -이때, 해설문은 제시된 정답이 옳은지를를 중심으로 문제의 목적을 통해 정답을 유도할 수 있도록 구성하고, 오답 설계 포인트와 혼동 지점을 반드시 제시하라.\n                        
-                        -해설문의 결론은 무조건 '따라서 정답은 ~이다.'로 구성하여라.\n
-                        -반드시 텍스트는 모든 일반 텍스트로 처리하고, '[]'를 제외한 마크다운 형식의 표현은 절대 사용하지 마라. 특히, '---'과 같은 구분선을 절대 출력하지 마라.\n
-                        -출력 문장의 종결 어미는 반드시 '~한다', '~이다' 등 비격식체 서술형 어미로 통일하라. 절대'-합니다', '-입니다' 등 격식체 종결어미는 절대 사용하지 말고, 문체는 일관되게 보고서식 비격식체로 유지하라. 
-                        \n\n\n{content}'''
-        },
-        "문제 생성_서술형": {
-            "system": f"너는 {field}에서 20년 경력을 지닌 평가 설계 전문가로, {question_level} 수준의 학습자를 위한 논리적 흐름, 절차적 지식의 이해를 평가하는 서술형 문항을 설계하는 데 특화되어 있다.",
-            "user": f'''학습자가 핵심 개념, 절차, 원리를 자신의 언어로 정확하고 일관되게 설명하는 능력을 평가할 수 있도록 서술형 문제를 구성하라.\n
-                        문제 생성 형식은 다음을 따르라.\n\n
-
-                        [비전공자 예시]\n
-                        문제1. 인조 나뭇잎이 물을 움직일 수 있는 이유를 자연의 원리와 연결하여 서술하시오.\n
-                        [정답]
-                        인조 나뭇잎은 자연의 잎 구조를 모방하여 물이 자연스럽게 이동할 수 있도록 만든 장치이다. 
-                        잎의 내부에는 미세한 통로들이 있어 물이 모세관 현상이나 표면장력 등의 물리적 원리를 통해 움직이는데, 인조 나뭇잎도 비슷하게 내부 구조를 형성하여 물이 아래에서 위로 이동할 수 있게 한다. 이 장치는 외부 에너지 없이도 물을 이동시키므로 ‘무동력’ 펌프라고 부른다.\n
-                        [해설]
-                        이 문항은 자연 원리(모세관 현상, 표면장력 등)가 인조 나뭇잎에 어떻게 적용되는지를 학습자가 이해하고 자신의 언어로 설명할 수 있는지를 평가한다.
-                        핵심은 “물리적 원리 기반의 구조적 모사”이며, 단순한 설명이 아니라 ‘왜 그런 구조가 가능한지’, ‘자연의 어떤 원리를 응용했는지’를 언급하는 것이 중요하다.
-                        자주 실수하는 부분은 단지 '물을 흡수한다'는 결과만 서술하고 구조나 원리를 설명하지 않는 것이다.
-                        이 문항의 채점 기준은 다음과 같다.\n
-                        [채점 기준]
-                        -자연의 물 이동 원리(예: 모세관, 표면장력 등)를 언급했는가
-                        -인조 나뭇잎이 이를 어떻게 모방하는지를 설명했는가
-                        -외부 에너지 없이 물이 이동하는 구조적 특징을 서술했는가\n\n
-
-                        [전공자 예시]\n
-
-                        문제1. cryo-gel 기법이 인조 나뭇잎의 수분 이송 성능을 향상시키는 구조적 원리를 중심으로 설명하시오.\n
-                        [정답]
-                        cryo-gel 기법은 저온 상태에서 겔을 형성하는 방식으로, 인조 나뭇잎 내부에 미세한 채널 구조를 만들어 수분이 통과할 수 있는 유로를 제공한다. 이러한 구조는 물이 흐를 때 발생하는 유동 저항을 감소시켜 수분의 이송 속도와 효율을 높이는 데 기여한다.
-                        또한, cryo-gel 기법은 구조적으로 다공성을 확보하기 때문에, 겔의 내부에서 일어나는 물의 모세관 흐름을 가능하게 하며, 이로 인해 무동력 상태에서도 안정적인 물 이송이 가능해진다.\n
-                        [해설]
-                        이 문항은 cryo-gel 기법의 구조 형성 메커니즘과 그에 따른 물리적 효과(유동 저항 감소, 유로 형성 등)를 이해하고 설명할 수 있는지를 평가한다.\n
-                        정답은 단순히 cryo-gel의 정의를 나열하는 것이 아니라, 해당 기법이 인조 나뭇잎 구조 내에서 어떤 기능을 수행하며 성능 향상에 어떻게 연결되는지를 명확히 설명해야 한다.
-                        자주 실수하는 부분은 종종 '겔 형성 기술'에만 초점을 두거나, '유동 저항 감소'를 구조와 연결짓지 못하는 데서 발생한다.
-                        이 문항의 채점 기준은 다음과 같다.\n
-                        [채점 기준]
-                        -cryo-gel 기법의 형성 방식(저온 겔화, 미세 채널 형성)을 정확히 설명했는가
-                        -유동 저항 감소와의 인과 관계를 구조적 관점에서 서술했는가
-                        -해당 구조가 인조 나뭇잎의 성능 향상에 어떤 영향을 주는지를 기술했는가\n\n
-                        -문제의 내용은 반드시 요약 결과를 바탕으로 구성하고, {field_features(field)}의 관점을 유지하여 {question_level} 학습자의 사고 수준에 따라 문제를 반드시 다음 중 하나의 SOLO 단계에 맞게 구성하라.\n
-                        -이때, 학습자의 수준이 {question_level}인 경우, 반드시 다음과 같은 사고 구조에 기반하여 문항을 설계하라: {user_question_level(question_level)}\n
-                        -또한, {question_level}이 비전공자인 경우, SOLO의 1-3단계 구조로, 전공자인 경우, SOLO의 4-6단계 구조로 문제를 구성하라. 반드시 여러 단계를 적절히 활용하여 문제를 구성하라.\n
-                        -문제의 수는 반드시 {question_count}개만 생성하라.\n
-                        -지시문의 문장 구조는 "~의 개념과 주요 특징을 서술하시오.?", "~에 대한 자신의 견해를 서술하시오.", "~의 차이점을 서술하시오?" 등 예시의 형식으로만 고정하지 말고, 반드시 다양한 구조를 활용하라.\n
-                        -정답은 학습자가 자신의 응답과 비교하여 핵심 개념 이해, 논리적 설명력, 구조화 수준을 점검할 수 있도록 구성된 피드백형 예시로 구성하라.\n
-                        -해설문은 다음의 기준을 모두 포함하여 작성하라.\n
-                          1. 문제에서 요구한 설명의 범위(개념, 사례, 비교 등)를 간략히 요약하고, 평가의 핵심 목적을 제시하라.\n
-                          2. 정답 서술이 반드시 포함해야 할 핵심 개념 또는 필수 요소를 명확히 정리하라.\n
-                          3. 해당 개념이 왜 중요한지, 어떤 원리나 사례와 연결되는지를 서술하라.\n
-                          4. 자주 발생하는 오개념 또는 학습자의 서술상 오류를 제시하고, 이를 구분할 수 있도록 설명하라.\n
-                          5. 학습자가 자기 답안을 점검할 수 있도록 채점 기준을 항목별로 간단하게 제시하라. 
-                          6. 해설은 정답의 근거, 관련 개념 간의 연결, 오개념과의 구분 등을 포함해 학습 효과를 높이도록 구성하라.\n\n
-                        -반드시 텍스트는 모든 일반 텍스트로 처리하고, '[]'를 제외한 마크다운 형식의 표현은 절대 사용하지 마라. 특히, '---'과 같은 구분선을 절대 출력하지 마라.\n
-                        -출력 문장의 종결 어미는 반드시 '~한다', '~이다' 등 비격식체 서술형 어미로 통일하라. 절대'-합니다', '-입니다' 등 격식체 종결어미는 절대 사용하지 말고, 문체는 일관되게 보고서식 비격식체로 유지하라. 
-                        -반드시 '[해설]', '[정답]'의 형태를 갖추어 출력하라. 절대 '[]'기호 없이 '해설', '정답'를 출력해서는 안 된다.\n 
-                        \n\n\n{content}'''
-        }
-    }
-
-    return prompts.get(type_name, {"system": "", "user": ""})
+__all__ = [
+    'get_base_url',
+    'truncate_by_tokens', 
+    'summarize_with_chatgpt',
+    'extract_text_from_pdf',
+    'extract_text_from_pptx',
+    
+    'get_summary_prompt'
+]
